### 2024-09-02 #7166 @editorialbot
[PRE REVIEW]: ARCHES PiCar-X: Software for Digital Twin Research
**Submitting author:** @AlexanderBarbie (<a href="http://orcid.org/0000-0001-5352-6039">Alexander Barbie</a>)
**Repository:** https://github.com/cau-se/ARCHES-PiCar-X
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
