### 2024-09-02 #7166 @AlexanderBarbie
Thank you, great list! Here are 5 potential reviewers with different backgrounds:

[@nuest](https://reviewers.joss.theoj.org/reviewers/40) -- This author is from the list above
[Martin Huber](https://reviewers.joss.theoj.org/reviewers/3026)
[Christopher Thierauf](https://reviewers.joss.theoj.org/reviewers/3180)
[Alexander Fabisch](https://reviewers.joss.theoj.org/reviewers/4024)
[Emerson Martins de Andrade](https://reviewers.joss.theoj.org/reviewers/3241)
