### 2024-09-03 #7166 @danielskatz
👋 @cst0, @AlexanderFabisch, @mrsonandrade - would one or more of you be willing to review this submission for JOSS?
