### 2024-09-03 #7166 @mrsonandrade
Hi @danielskatz! I am available to review this submission for JOSS.
