### 2024-09-05 #7179 @AlexanderFabisch
@AlexanderBarbie: I have a question regarding this point of the checklist:

> Contribution and authorship: Has the submitting author (@AlexanderBarbie) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

According to the list above, your co-author Wilhelm Hasselbring only made one commit to the repositoy

>     41	Alexander Barbie
>     1	AlexanderBarbie
>     1	Wilhelm Hasselbring
>     1	abarbie

I think this would hardly qualify as an appropriate contribution to claim co-authorship. However, comparing the amount of code and number of commits, I assume that you developed the project internally and pushed it to github without the whole commit history. Can you confirm this? Did your co-author contribute a significant part to the software?
