### 2024-09-05 #7179 @danielskatz
@AlexanderFabisch - please keep in mind that JOSS is looking for authors to have contributed to the project, which can happen in many ways outside of commits to a repo.
