### 2024-09-05 cau-se/ARCHES-PiCar-X#3 @AlexanderFabisch
Incorrect path in installation instructions
I just copied this part of the [installation instructions](https://cau-se.github.io/ARCHES-PiCar-X/getting_started.html):

    git clone https://github.com/cau-se/ARCHES-PiCar-X.git
    cd ./ARCHES-Picar-X/PiCar-X

does not work for me. It was a bit hard to figure out why, but the second c in the path should be upper case:

    cd ./ARCHES-PiCar-X/PiCar-X

openjournals/joss-reviews#7179
