### 2024-09-05 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
Error in documentation "Getting Started" 
**Describe the bug**

When starting the docker containers, I get the following error messages:

```
picarx-gazebo-1             | [gazebo_gui-3] process has died [pid 52, exit code 134, cmd /opt/ros/noetic/lib/gazebo_ros/gzclient __name:=gazebo_gui __log:=/root/.ros/log/884fc2d4-6b84-11ef-afb6-0242ac120003/gazebo_gui-3.log].

picarx-gazebo-1             | [ERROR] [1725540245.266274479, 0.175000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_right

picarx-gazebo-1             | [ERROR] [1725540245.266727800, 0.175000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_left

picarx-gazebo-1             | [ERROR] [1725540245.267164615, 0.175000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_left

picarx-gazebo-1             | [ERROR] [1725540245.267614366, 0.175000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_right

motor_right-dtp-1           | [motor_right-1] process has died [pid 32, exit code 1, cmd /root/catkin_ws/install/lib/picarx_dcmotor_driver/dcmotor_driver.py motor_right 23 P13 /dev/i2c-0 1 __name:=motor_right __log:=/root/.ros/log/884bf078-6b84-11ef-a227-0242ac120002/motor_right-1.log].

steering-dtp-1              | [ackermann_clutchgear_driver-1] process has died [pid 32, exit code 1, cmd /root/catkin_ws/install/lib/picarx_clutchgear_driver/ackermann_clutchgear.py ackermann_clutchgear_driver P2 /dev/i2c-0 __name:=ackermann_clutchgear_driver __log:=/root/.ros/log/884bf078-6b84-11ef-a227-0242ac120002/ackermann_clutchgear_driver-1.log].

steering_emulator-dtp-1     | [ackermann_clutchgear_emulator-1] process has died [pid 32, exit code 1, cmd /root/catkin_ws/install/lib/picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py ackermann_clutchgear_emulator P2 /dev/i2c-0 __name:=ackermann_clutchgear_emulator __log:=/root/.ros/log/884fc2d4-6b84-11ef-afb6-0242ac120003/ackermann_clutchgear_emulator-1.log].

motor_left-dtp-1            | [motor_left-1] process has died [pid 32, exit code 1, cmd /root/catkin_ws/install/lib/picarx_dcmotor_driver/dcmotor_driver.py motor_left 24 P12 /dev/i2c-0 0 __name:=motor_left __log:=/root/.ros/log/884bf078-6b84-11ef-a227-0242ac120002/motor_left-1.log].

mqtt-dtp-1                  | [MQTTTwinBridge-1] process has died [pid 31, exit code 1, cmd /root/catkin_ws/install/lib/arches_mqtt_bridge/twin_bridge_pt.py MQTTTwinBridge None __name:=MQTTTwinBridge __log:=/root/.ros/log/884bf078-6b84-11ef-a227-0242ac120002/MQTTTwinBridge-1.log].

 exit code 1, cmd /root/catkin_ws/install/lib/picarx_dcmotor_emulator/dcmotor_emulator.py motor_emulator_left 24 P12 /dev/i2c-0 0 __name:=motor_emulator_left __log:=/root/.ros/log/884fc2d4-6b84-11ef-afb6-0242ac120003/motor_emulator_left-1.log].
```

**To Reproduce**
Steps to reproduce the behavior:
1. Follow the [Quick Start](https://cau-se.github.io/ARCHES-PiCar-X/getting_started.html) guide
    1. Compile linux kernel
    2. Activate GPIO and I2C on System
    3. Build and Start Docker Containers from Scratch
    4. Run `docker compose -f docker-compose-dtp.yml up`

 - OS: Ubuntu 20.04, Linux kernel 5.13

openjournals/joss-reviews/issues/7179
