### 2024-09-05 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
Can you please provide more information on the error, for instance in the motor_right-dtp-1 container? There should be a ROS error message, if the process crashes.
