### 2024-09-06 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
When I run `docker compose -f docker-compose-dtp.yml up` I tried to attach to `picar-x-motor_emulator_right-dtp-1` as quickly as possible. This is the output:

```
docker attach picar-x-motor_emulator_right-dtp-1

... logging to /root/.ros/log/70b0838c-6c26-11ef-b4e9-0242ac120002/roslaunch-b60bf20f98df-1.log
Checking log directory for disk usage. This may take a while.
Press Ctrl-C to interrupt
Done checking log file disk usage. Usage is <1GB.

started roslaunch server http://b60bf20f98df:40169/

SUMMARY
========

PARAMETERS
 * /motor_emulator_right/controller_publisher_topic: /picarx/rear_righ...
 * /motor_emulator_right/direction_pin: 23
 * /motor_emulator_right/i2c_port: /dev/i2c-0
 * /motor_emulator_right/motor_side: 1
 * /motor_emulator_right/name: motor_emulator_right
 * /motor_emulator_right/pwm_pin: P13
 * /rosdistro: noetic
 * /rosversion: 1.16.0

NODES
  /
    motor_emulator_right (picarx_dcmotor_emulator/dcmotor_emulator.py)

ROS_MASTER_URI=http://picarx-gazebo:11312

process[motor_emulator_right-1]: started with pid [31]
```

So there is no error here.

~~... but I noticed that I got the following error on the other terminal:~~

```
steering_emulator-dtp-1     | Traceback (most recent call last):
steering_emulator-dtp-1     |   File "/root/catkin_ws/install/lib/picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py", line 191, in <module>
steering_emulator-dtp-1     |     clutchgear_emulator = AckermannClutchGearEmulator(
steering_emulator-dtp-1     |   File "/root/catkin_ws/install/lib/picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py", line 67, in __init__
steering_emulator-dtp-1     |     super(AckermannClutchGearEmulator, self).__init__(pwm_pin, i2c_port)
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/emulators/clutchgear.py", line 27, in __init__
steering_emulator-dtp-1     |     self.pwm_pin = {'channel': pwm_pin, 'i2c_port': i2c_port}
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/emulators/clutchgear.py", line 46, in pwm_pin
steering_emulator-dtp-1     |     self.__pwm_pin = PWM(
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 81, in __init__
steering_emulator-dtp-1     |     super(PWM, self).__init__(i2c_port=i2c_port, address=address)
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/i2c.py", line 50, in __init__
steering_emulator-dtp-1     |     self.smbus = SMBus(self.i2c_port)
steering_emulator-dtp-1     |   File "/usr/local/lib/python3.8/dist-packages/smbus2/smbus2.py", line 280, in __init__
steering_emulator-dtp-1     |     self.open(bus)
steering_emulator-dtp-1     |   File "/usr/local/lib/python3.8/dist-packages/smbus2/smbus2.py", line 310, in open
steering_emulator-dtp-1     |     self.fd = os.open(filepath, os.O_RDWR)
steering_emulator-dtp-1     | IsADirectoryError: [Errno 21] Is a directory: '/dev/i2c-0'
```

~~I assume it is mounted from the host into the container. I also checked it on the host system and this is correct, `/dev/i2c-0` is a directory. Is that a problem?~~

**edit: I found [this section](https://cau-se.github.io/ARCHES-PiCar-X/troubleshooting.html#activate-gpio-and-i2c-on-your-system) of the documentation and fixed this issue.**

On the same terminal I got

```
motor_left-dtp-1            | OSError: [Errno 22] Invalid argument
motor_left-dtp-1            |
motor_left-dtp-1            | During handling of the above exception, another exception occurred:
motor_left-dtp-1            |
motor_left-dtp-1            | Traceback (most recent call last):
motor_left-dtp-1            |   File "/root/catkin_ws/install/lib/picarx_dcmotor_driver/dcmotor_driver.py", line 137, in <module>
motor_left-dtp-1            |     dcmotor = DCMotorDriver(options.args.name, options.args.direction_pin,
motor_left-dtp-1            |   File "/root/catkin_ws/install/lib/picarx_dcmotor_driver/dcmotor_driver.py", line 91, in __init__
motor_left-dtp-1            |     super(DCMotorDriver, self).__init__(name, int(direction_pin),
motor_left-dtp-1            |   File "/root/catkin_ws/src/core/picarx/src/picarx/drivers/dcmotor.py", line 25, in __init__
motor_left-dtp-1            |     self.direction_pin = direction_pin
motor_left-dtp-1            |   File "/root/catkin_ws/src/core/picarx/src/picarx/drivers/dcmotor.py", line 65, in direction_pin
motor_left-dtp-1            |     self.__direction_pin = GPIO().setup(
motor_left-dtp-1            |   File "/root/catkin_ws/src/core/picarx/src/picarx/gpio.py", line 407, in setup
motor_left-dtp-1            |     new_pin = Pin(pin_number, direction=direction, callback=callback)
motor_left-dtp-1            |   File "/root/catkin_ws/src/core/picarx/src/picarx/gpio.py", line 141, in __init__
motor_left-dtp-1            |     self.pin_number = pin_number
motor_left-dtp-1            |   File "/root/catkin_ws/src/core/picarx/src/picarx/gpio.py", line 196, in pin_number
motor_left-dtp-1            |     self.__export()
motor_left-dtp-1            |   File "/root/catkin_ws/src/core/picarx/src/picarx/gpio.py", line 305, in __export
motor_left-dtp-1            |     export.write('%d' % self.pin_number)
motor_left-dtp-1            | OSError: [Errno 22] Invalid argument
```

and

```
mqtt-dtp-1                  | Traceback (most recent call last):
mqtt-dtp-1                  |   File "/root/catkin_ws/install/lib/arches_mqtt_bridge/twin_bridge_pt.py", line 19, in <module>
mqtt-dtp-1                  |     twin_bridge.start()
mqtt-dtp-1                  |   File "/root/catkin_ws/install/lib/python3/dist-packages/arches_mqtt_bridge/twin_bridge.py", line 106, in start
mqtt-dtp-1                  |     self.client = rospy.get_param("~config")
mqtt-dtp-1                  |   File "/root/catkin_ws/install/lib/python3/dist-packages/arches_mqtt_bridge/twin_bridge.py", line 50, in client
mqtt-dtp-1                  |     self.__client = MQTTClientFactory.create_client(
mqtt-dtp-1                  |   File "/root/catkin_ws/install/lib/python3/dist-packages/arches_mqtt_bridge/client.py", line 14, in create_client
mqtt-dtp-1                  |     cls.__connection(client, config['connect'])
mqtt-dtp-1                  |   File "/root/catkin_ws/install/lib/python3/dist-packages/arches_mqtt_bridge/client.py", line 49, in __connection
mqtt-dtp-1                  |     client.connect(**connection_config)
mqtt-dtp-1                  |   File "/usr/local/lib/python3.8/dist-packages/paho/mqtt/client.py", line 1435, in connect
mqtt-dtp-1                  |     return self.reconnect()
mqtt-dtp-1                  |   File "/usr/local/lib/python3.8/dist-packages/paho/mqtt/client.py", line 1598, in reconnect
mqtt-dtp-1                  |     self._sock = self._create_socket()
mqtt-dtp-1                  |   File "/usr/local/lib/python3.8/dist-packages/paho/mqtt/client.py", line 4609, in _create_socket
mqtt-dtp-1                  |     sock = self._create_socket_connection()
mqtt-dtp-1                  |   File "/usr/local/lib/python3.8/dist-packages/paho/mqtt/client.py", line 4640, in _create_socket_connection
mqtt-dtp-1                  |     return socket.create_connection(addr, timeout=self._connect_timeout, source_address=source)
mqtt-dtp-1                  |   File "/usr/lib/python3.8/socket.py", line 787, in create_connection
mqtt-dtp-1                  |     for res in getaddrinfo(host, port, 0, SOCK_STREAM):
mqtt-dtp-1                  |   File "/usr/lib/python3.8/socket.py", line 918, in getaddrinfo
mqtt-dtp-1                  |     for res in _socket.getaddrinfo(host, port, family, type, proto, flags):
mqtt-dtp-1                  | socket.gaierror: [Errno -3] Temporary failure in name resolution
```
