### 2024-09-06 cau-se/ARCHES-PiCar-X#5 @AlexanderFabisch
Documentation
* I'd suggest to link the documentation (https://cau-se.github.io/ARCHES-PiCar-X/) directly in the project description (right top: project website).
* When I click on "Get started now" I get a 404 error.

openjournals/joss-reviews/issues/7179
