### 2024-09-06 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
Okay, the I2C error was the problem with the folder, removing it helped.

I assume the error on the motor_left-dtp-1 is now a problem with the gpio pins:
```console
File "/root/catkin_ws/src/core/picarx/src/picarx/gpio.py", line 305, in __export
motor_left-dtp-1            |     export.write('%d' % self.pin_number)
motor_left-dtp-1            | OSError: [Errno 22] Invalid argument
```

If you shut down the containers with
```console
docker compose down
````

and then execute
```console
ls /sys/class/gpio
```

You should get the result:
```console
export  gpiochip1  unexport
```

if there are gpio pins `gpio23` and `gpio24` listed, then you have to remove them. I think the easierst way is just a restart and then you activate the three modules again. If this does not delete the gpio pins, then execute:

```console
echo 24 > /sys/class/gpio/unexport
echo 23 > /sys/class/gpio/unexport
```

I am trying to replicate this error on a fresh Ubuntu 20.04.

Edit: You can ignore the mqtt error, if you are not trying to connect the dtp with a running mqtt server.
