### 2024-09-06 8ed9583
fixes #3
fixes the wrong path to the cloned folder
