### 2024-09-06 cau-se/ARCHES-PiCar-X#5 @AlexanderBarbie
sorry, first time using the github pages stuff and changed the structure several times.
