### 2024-09-06 #7179 @AlexanderBarbie
> @AlexanderBarbie: I have a question regarding this point of the checklist:
>
> > Contribution and authorship: Has the submitting author (@AlexanderBarbie) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> According to the list above, your co-author Wilhelm Hasselbring only made one commit to the repositoy
>
> > ```
> > 41	Alexander Barbie
> > 1	AlexanderBarbie
> > 1	Wilhelm Hasselbring
> > 1	abarbie
> > ```
>
> I think this would hardly qualify as an appropriate contribution to claim co-authorship. However, comparing the amount of code and number of commits, I assume that you developed the project internally and pushed it to github without the whole commit history. Can you confirm this? Or did your co-author contribute a significant part to the ~software~ project (maybe in any other form)?

@whasselbring and I have been conducting research together for six years on how digital twins, particularly digital twin prototypes, can be developed and applied in robotics. After proving the concept that DTPs work, we collaboratively developed a formal specification of our digital twin concept.

PoC: `Barbie, A., Pech, N., Hasselbring, W., Flögel, S., Wenzhöfer, F., Walter, M., … & Sommer, S. (2021). Developing an underwater network of ocean observation systems with digital twin prototypes—a field report from the baltic sea. IEEE Internet Computing, 26(3), 33-42. https://doi.org/10.1109/mic.2021.3065245`

Formalization: `Barbie, A., & Hasselbring, W. (2024). From Digital Twins to Digital Twin Prototypes: Concepts, Formalization, and Applications. IEEE Access. https://doi.org/10.1109/access.2024.3406510`

It was his idea to create a low-cost lab experiment to easily demonstrate the results and allow the replication of experiments for other researchers. The PiCar-X emerged from this idea and is used in another publication to illustrate the formal specification with practical examples:

PiCar-X Applied: `Barbie, A., & Hasselbring, W. (2024). Toward Reproducibility of Digital Twin Research: Exemplified with the PiCar-X. arXiv preprint arXiv:2408.13866. https://doi.org/10.48550/ARXIV.2408.13866`

While his direct contribution to the source code may not be extensive, his work is an integral part of the overall concept, and without it, this submission would not have been realized.
