### 2024-09-06 cau-se/ARCHES-PiCar-X#5 @AlexanderFabisch
> I'd suggest to link the documentation (https://cau-se.github.io/ARCHES-PiCar-X/) directly in the project description (right top: project website).

What about this? Just a recommendation, but you didn't comment on it.
