### 2024-09-06 #7179 @AlexanderFabisch
> While his direct contribution to the source code may not be extensive, his work is an integral part of the overall concept, and without it, this submission would not have been realized.

OK, thanks for clarification.
