### 2024-09-06 cau-se/ARCHES-PiCar-X#5 @AlexanderBarbie
had to google it first. :)
was wondering all the time how to link the github pages on the front page of the repo
