### 2024-09-08 cab8e32
added I2C variable to env and adjusted docs #4
