### 2024-09-08 af62e8b
setting default i2c value in simulation.env to 0 for tests
