+++ .github/workflows/push-x64.yml
@@ -54,7 +54,7 @@ jobs:
       - name: Speed tests using Gazebo headless
         working-directory: ./PiCar-X
         run: |
-          i2c=/dev/i2c-0 docker compose -f docker-compose-dtp-headless-gazebo.yml up -d
+          I2C=/dev/i2c-0 docker compose -f docker-compose-dtp-headless-gazebo.yml up -d
           docker exec picar-x-picarx-gazebo-control-1 /bin/bash -c "source ./install/setup.bash && sleep 30; python3 ./src/simulation/picarx_control/tests/steering_integration_test.py"
 
       - name: Remove running containers after tests
