### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
I'll just post my bash history here:

```
$ docker compose -f docker-compose-dtp.yml down
[+] Running 13/13
 ✔ Container picar-x-picarx-gazebo-control-1     Removed                   0.0s
 ✔ Container picar-x-motor_right-dtp-1           Removed                   0.0s
 ✔ Container picar-x-motor_left-dtp-1            Removed                   0.0s
 ✔ Container picar-x-shadow-dtp-1                Removed                   0.0s
 ✔ Container picar-x-ackermann_skill-dtp-1       Removed                   0.0s
 ✔ Container picar-x-motor_emulator_left-dtp-1   Removed                   0.0s
 ✔ Container picar-x-mqtt-dtp-1                  Removed                   0.0s
 ✔ Container picar-x-motor_emulator_right-dtp-1  Removed                   0.0s
 ✔ Container picar-x-steering_emulator-dtp-1     Removed                   0.0s
 ✔ Container picar-x-steering-dtp-1              Removed                   0.0s
 ✔ Container picar-x-picarx-1                    Removed                   0.0s
 ✔ Container picar-x-picarx-gazebo-1             Removed                   0.0s
 ✔ Network picar-x_default                       Removed                   0.1s

$ ls /sys/class/gpio
export  gpiochip81  unexport

echo 81 > /sys/class/gpio/unexport
bash: /sys/class/gpio/unexport: Permission denied
```

Switching into a root bash with sudo results in

```
bash: echo: write error: Invalid argument
```

> I think the easierst way is just a restart and then you activate the three modules again. If this does not delete the gpio pins, then execute:

What do you mean by restarting? The docker? The OS? The module? Now I seem to have more GPIO pins: :D

```
$ ls /sys/class/gpio
export  gpiochip1  gpiochip81  unexport
```

I'm a bit lost here. I'm not very familiar with GPIO stuff.
