### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
Sorry, looks like I did not send the comment I wrote after the commit, will check if I am home.

The log looks fine, no active gpio pin.

You could try the following:

Create a file, e.g.:

```console
nano export-gpio.bash
```

Enter in that file

```console
echo 23 > /sys/class/gpio/export
echo 24 > /sys/class/gpio/export
```

Save and execute:

```console
chmod +x export-gpio.bash
```

And then execute with `sudo export-gpio.bash`.

With `ls /sys/class/gpio/*` you should now see something like `gpio23` and `gpio24`.

To unexport the pins again, create a file with the content:

```console
echo 23 > /sys/class/gpio/unexport
echo 24 > /sys/class/gpio/unexport
```

I already added this to the documentation yesterday.
