### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
So I got this:

```
$ ls /sys/class/gpio/*
/sys/class/gpio/export  /sys/class/gpio/unexport

/sys/class/gpio/gpio23:
active_low  device  direction  edge  power  subsystem  uevent  value

/sys/class/gpio/gpio24:
active_low  device  direction  edge  power  subsystem  uevent  value

/sys/class/gpio/gpiochip1:
base  device  label  ngpio  power  subsystem  uevent

/sys/class/gpio/gpiochip81:
base  device  label  ngpio  power  subsystem  uevent
```

So the following does not work directly. I have to put it in a script and run it with `sudo ./script.sh` (which works for the previously exported 23 and 24).

```
echo 23 > /sys/class/gpio/unexport
echo 24 > /sys/class/gpio/unexport
```

But when I try to unexport 1 or 81, it fails with

```
$ sudo ./unexport-gpio.bash
./unexport-gpio.bash: 1: echo: echo: I/O error
./unexport-gpio.bash: 2: echo: echo: I/O error
```
