### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
Okay, then the gpio modules are installed correctly.

I suggest the following:

1. reboot once
2. Check with `ls /dev/i2c*` if there is already an i2c available.
There could be `/dev/i2c-0` by default.
3. activate the three modules with modprobe
4. Check if there is a new i2c available, e.g. `/dev/i2c-1`
5. Set the I2C variable in `ARCHES-PiCar-X/PiCar-X/env/simulation.env` to the new I2C device.
6. Execute in den `PiCaR-X` folder `docker compose -f docker-compose-dtp.yml up`.
7. Paste again the log from a crashing container (not mqtt)
