### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
I'll just show the results here:

```
$ ls /dev/i2c*
ls: cannot access '/dev/i2c*': No such file or directory

$ sudo modprobe gpio-mockup gpio_mockup_ranges=1,41
$ sudo modprobe i2c-dev
$ sudo modprobe i2c-stub chip_addr=0x14

$ ls /dev/i2c*
/dev/i2c-0  /dev/i2c-1  /dev/i2c-2  /dev/i2c-3
```

Now `ARCHES-PiCar-X/PiCar-X/env/simulation.env` only contains

```
ROS_MASTER_URI=http://picarx-gazebo:11312
ROS_HOST_PORT=11312
```

~~Which I2C variable should I set? And to which value exactly?~~

OK, I updated to the latest commit. Now it contains

```
ROS_MASTER_URI=http://picarx-gazebo:11312
ROS_HOST_PORT=11312
I2C=/dev/i2c-0
```

I believe that is OK? Or should I choose a different one?

<summary>Click to show the full log.</summary>
<pre>
docker compose -f docker-compose-dtp.yml up
[+] Running 13/13
 ✔ Network picar-x_default                       Created                   0.0s
 ✔ Container picar-x-picarx-1                    Created                   0.1s
 ✔ Container picar-x-picarx-gazebo-1             Created                   0.1s
 ✔ Container picar-x-mqtt-dtp-1                  Created                   0.0s
 ✔ Container picar-x-ackermann_skill-dtp-1       Created                   0.0s
 ✔ Container picar-x-motor_right-dtp-1           Created                   0.0s
 ✔ Container picar-x-motor_left-dtp-1            Created                   0.0s
 ✔ Container picar-x-shadow-dtp-1                Created                   0.0s
 ✔ Container picar-x-steering-dtp-1              Created                   0.0s
 ✔ Container picar-x-picarx-gazebo-control-1     Created                   0.0s
 ✔ Container picar-x-steering_emulator-dtp-1     Created                   0.0s
 ✔ Container picar-x-motor_emulator_left-dtp-1   Created                   0.0s
 ✔ Container picar-x-motor_emulator_right-dtp-1  Created                   0.0s
Attaching to ackermann_skill-dtp-1, motor_emulator_left-dtp-1, motor_emulator_right-dtp-1, motor_left-dtp-1, motor_right-dtp-1, mqtt-dtp-1, picarx-1, picarx-gazebo-1, picarx-gazebo-control-1, shadow-dtp-1, steering-dtp-1, steering_emulator-dtp-1
picarx-1                    | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-e30d2f51b567-1.log
picarx-1                    | Checking log directory for disk usage. This may take a while.
picarx-1                    | Press Ctrl-C to interrupt
picarx-1                    | Done checking log file disk usage. Usage is <1GB.
picarx-1                    |
picarx-gazebo-1             | ... logging to /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/roslaunch-123d75246fcb-1.log
picarx-gazebo-1             | Checking log directory for disk usage. This may take a while.
picarx-gazebo-1             | Press Ctrl-C to interrupt
picarx-gazebo-1             | Done checking log file disk usage. Usage is <1GB.
picarx-gazebo-1             |
picarx-1                    | started roslaunch server http://picarx:34043/
picarx-1                    | ros_comm version 1.16.0
picarx-1                    |
picarx-1                    |
picarx-1                    | SUMMARY
picarx-1                    | ========
picarx-1                    |
picarx-1                    | PARAMETERS
picarx-1                    |  * /rosdistro: noetic
picarx-1                    |  * /rosversion: 1.16.0
picarx-1                    |
picarx-1                    | NODES
picarx-1                    |
picarx-1                    | auto-starting new master
picarx-1                    | process[master]: started with pid [32]
picarx-gazebo-1             | started roslaunch server http://picarx-gazebo:33733/
picarx-gazebo-1             |
picarx-gazebo-1             | SUMMARY
picarx-gazebo-1             | ========
picarx-gazebo-1             |
picarx-gazebo-1             | PARAMETERS
picarx-gazebo-1             |  * /gazebo/enable_ros_network: True
picarx-gazebo-1             |  * /robot_description: <?xml version="1....
picarx-gazebo-1             |  * /rosdistro: noetic
picarx-gazebo-1             |  * /rosversion: 1.16.0
picarx-gazebo-1             |  * /use_sim_time: True
picarx-gazebo-1             |
picarx-gazebo-1             | NODES
picarx-gazebo-1             |   /
picarx-gazebo-1             |     gazebo (gazebo_ros/gzserver)
picarx-gazebo-1             |     gazebo_gui (gazebo_ros/gzclient)
picarx-gazebo-1             |     spawn_urdf (gazebo_ros/spawn_model)
picarx-gazebo-1             |
picarx-gazebo-1             | auto-starting new master
picarx-gazebo-1             | process[master]: started with pid [34]
picarx-1                    | ROS_MASTER_URI=http://picarx:11311/
picarx-1                    |
picarx-1                    | setting /run_id to 0924a26a-6ec0-11ef-80d4-0242ac120003
picarx-1                    | process[rosout-1]: started with pid [42]
picarx-1                    | started core service [/rosout]
picarx-gazebo-1             | ROS_MASTER_URI=http://picarx-gazebo:11312
picarx-gazebo-1             |
picarx-gazebo-1             | setting /run_id to 09314c90-6ec0-11ef-84d9-0242ac120002
picarx-gazebo-1             | process[rosout-1]: started with pid [44]
picarx-gazebo-1             | started core service [/rosout]
picarx-gazebo-1             | process[gazebo-2]: started with pid [47]
picarx-gazebo-1             | process[gazebo_gui-3]: started with pid [54]
picarx-gazebo-1             | process[spawn_urdf-4]: started with pid [61]
picarx-gazebo-1             | [ INFO] [1725895654.925857761]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1725895654.926633478]: waitForService: Service [/gazebo/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1725895655.033635241]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1725895655.034459549]: waitForService: Service [/gazebo_gui/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1725895655.090414086, 0.010000000]: waitForService: Service [/gazebo/set_physics_properties] is now available.
picarx-gazebo-1             | [ INFO] [1725895655.103525119, 0.023000000]: Physics dynamic reconfigure ready.
picarx-gazebo-1             | Aborted (core dumped)
picarx-gazebo-1             | [gazebo_gui-3] process has died [pid 54, exit code 134, cmd /opt/ros/noetic/lib/gazebo_ros/gzclient __name:=gazebo_gui __log:=/root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/gazebo_gui-3.log].
picarx-gazebo-1             | log file: /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/gazebo_gui-3*.log
picarx-gazebo-1             | [ INFO] [1725895655.585349892, 0.375000000]: Loading gazebo_ros_control plugin
picarx-gazebo-1             | [ INFO] [1725895655.585418570, 0.375000000]: Starting gazebo_ros_control plugin in namespace: /picarx
picarx-gazebo-1             | [ INFO] [1725895655.585793504, 0.375000000]: gazebo_ros_control plugin is waiting for model URDF in parameter [/robot_description] on the ROS param server.
picarx-gazebo-1             | [ERROR] [1725895655.698061576, 0.375000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_right
picarx-gazebo-1             | [ERROR] [1725895655.698692740, 0.375000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_left
picarx-gazebo-1             | [ERROR] [1725895655.699189398, 0.375000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_left
picarx-gazebo-1             | [ERROR] [1725895655.699564194, 0.375000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_right
picarx-gazebo-1             | [ INFO] [1725895655.703125045, 0.375000000]: Loaded gazebo_ros_control.
picarx-gazebo-1             | [spawn_urdf-4] process has finished cleanly
picarx-gazebo-1             | log file: /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/spawn_urdf-4*.log
steering_emulator-dtp-1     | ... logging to /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/roslaunch-f5398018f445-1.log
steering_emulator-dtp-1     | Checking log directory for disk usage. This may take a while.
steering_emulator-dtp-1     | Press Ctrl-C to interrupt
steering_emulator-dtp-1     | Done checking log file disk usage. Usage is <1GB.
steering_emulator-dtp-1     |
steering-dtp-1              | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-606bf7510fab-1.log
steering-dtp-1              | Checking log directory for disk usage. This may take a while.
steering-dtp-1              | Press Ctrl-C to interrupt
steering-dtp-1              | Done checking log file disk usage. Usage is <1GB.
steering-dtp-1              |
steering_emulator-dtp-1     | started roslaunch server http://f5398018f445:40647/
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | SUMMARY
steering_emulator-dtp-1     | ========
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | PARAMETERS
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/i2c_port: /dev/i2c-0
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/left_steer_topic: /picarx/steer_lef...
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/name: ackermann_clutchg...
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/pwm_pin: P2
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/right_steer_topic: /picarx/steer_rig...
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/wheel_base: 13.5
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/wheel_track: 9.2
steering_emulator-dtp-1     |  * /rosdistro: noetic
steering_emulator-dtp-1     |  * /rosversion: 1.16.0
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | NODES
steering_emulator-dtp-1     |   /
steering_emulator-dtp-1     |     ackermann_clutchgear_emulator (picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py)
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | ROS_MASTER_URI=http://picarx-gazebo:11312
steering_emulator-dtp-1     |
steering-dtp-1              | started roslaunch server http://606bf7510fab:36431/
steering-dtp-1              |
steering-dtp-1              | SUMMARY
steering-dtp-1              | ========
steering-dtp-1              |
steering-dtp-1              | PARAMETERS
steering-dtp-1              |  * /ackermann_clutchgear_driver/i2c_port: /dev/i2c-1
steering-dtp-1              |  * /ackermann_clutchgear_driver/name: ackermann_clutchg...
steering-dtp-1              |  * /ackermann_clutchgear_driver/pwm_pin: P2
steering-dtp-1              |  * /ackermann_clutchgear_driver/steering_status_topic: steering/status
steering-dtp-1              |  * /ackermann_clutchgear_driver/steering_topic: steering/angle
steering-dtp-1              |  * /rosdistro: noetic
steering-dtp-1              |  * /rosversion: 1.16.0
steering-dtp-1              |
steering-dtp-1              | NODES
steering-dtp-1              |   /
steering-dtp-1              |     ackermann_clutchgear_driver (picarx_clutchgear_driver/ackermann_clutchgear.py)
steering-dtp-1              |
steering-dtp-1              | ROS_MASTER_URI=http://picarx:11311
steering-dtp-1              |
steering-dtp-1              | process[ackermann_clutchgear_driver-1]: started with pid [32]
steering_emulator-dtp-1     | process[ackermann_clutchgear_emulator-1]: started with pid [31]
ackermann_skill-dtp-1       | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-76a6681eec60-1.log
ackermann_skill-dtp-1       | Checking log directory for disk usage. This may take a while.
ackermann_skill-dtp-1       | Press Ctrl-C to interrupt
ackermann_skill-dtp-1       | Done checking log file disk usage. Usage is <1GB.
ackermann_skill-dtp-1       |
steering-dtp-1              | Traceback (most recent call last):
steering-dtp-1              |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 121, in channel
steering-dtp-1              |     self.write_byte(0x2C)
steering-dtp-1              |   File "/root/catkin_ws/src/core/picarx/src/picarx/i2c.py", line 84, in write_byte
steering-dtp-1              |     return self.smbus.write_byte(self.address, data)
steering-dtp-1              |   File "/usr/local/lib/python3.8/dist-packages/smbus2/smbus2.py", line 414, in write_byte
steering-dtp-1              |     ioctl(self.fd, I2C_SMBUS, msg)
steering-dtp-1              | OSError: [Errno 121] Remote I/O error
steering-dtp-1              |
steering-dtp-1              | During handling of the above exception, another exception occurred:
steering-dtp-1              |
steering-dtp-1              | Traceback (most recent call last):
steering-dtp-1              |   File "/root/catkin_ws/install/lib/picarx_clutchgear_driver/ackermann_clutchgear.py", line 148, in <module>
steering-dtp-1              |     ackermann_clutchgear_driver = AckermannClutchGearDriver(
steering-dtp-1              |   File "/root/catkin_ws/install/lib/picarx_clutchgear_driver/ackermann_clutchgear.py", line 68, in __init__
steering-dtp-1              |     super(AckermannClutchGearDriver, self).__init__(pwm_pin, i2c_port)
steering-dtp-1              |   File "/root/catkin_ws/src/core/picarx/src/picarx/drivers/clutchgear.py", line 19, in __init__
steering-dtp-1              |     self.pwm_pin = {'channel': pwm_pin, 'i2c_port': i2c_port}
steering-dtp-1              |   File "/root/catkin_ws/src/core/picarx/src/picarx/drivers/clutchgear.py", line 38, in pwm_pin
steering-dtp-1              |     self.__pwm_pin = PWM(
steering-dtp-1              |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 83, in __init__
steering-dtp-1              |     self.channel = channel
steering-dtp-1              |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 125, in channel
steering-dtp-1              |     raise IOError("Can't send stuff.")
steering-dtp-1              | OSError: Can't send stuff.
motor_emulator_right-dtp-1  | ... logging to /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/roslaunch-e714c8e49335-1.log
motor_emulator_right-dtp-1  | Checking log directory for disk usage. This may take a while.
motor_emulator_right-dtp-1  | Press Ctrl-C to interrupt
motor_emulator_right-dtp-1  | Done checking log file disk usage. Usage is <1GB.
motor_emulator_right-dtp-1  |
motor_right-dtp-1           | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-b4d6f89d1b18-1.log
motor_right-dtp-1           | Checking log directory for disk usage. This may take a while.
motor_right-dtp-1           | Press Ctrl-C to interrupt
motor_right-dtp-1           | Done checking log file disk usage. Usage is <1GB.
motor_right-dtp-1           |
motor_emulator_left-dtp-1   | ... logging to /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/roslaunch-240f7c494817-1.log
motor_emulator_left-dtp-1   | Checking log directory for disk usage. This may take a while.
motor_emulator_left-dtp-1   | Press Ctrl-C to interrupt
motor_emulator_left-dtp-1   | Done checking log file disk usage. Usage is <1GB.
motor_emulator_left-dtp-1   |
motor_left-dtp-1            | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-faba2263b3a6-1.log
motor_left-dtp-1            | Checking log directory for disk usage. This may take a while.
motor_left-dtp-1            | Press Ctrl-C to interrupt
motor_left-dtp-1            | Done checking log file disk usage. Usage is <1GB.
motor_left-dtp-1            |
mqtt-dtp-1                  | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-6595a15394f0-1.log
mqtt-dtp-1                  | Checking log directory for disk usage. This may take a while.
mqtt-dtp-1                  | Press Ctrl-C to interrupt
mqtt-dtp-1                  | Done checking log file disk usage. Usage is <1GB.
mqtt-dtp-1                  |
steering_emulator-dtp-1     | Traceback (most recent call last):
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 121, in channel
steering_emulator-dtp-1     |     self.write_byte(0x2C)
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/i2c.py", line 84, in write_byte
steering_emulator-dtp-1     |     return self.smbus.write_byte(self.address, data)
steering_emulator-dtp-1     |   File "/usr/local/lib/python3.8/dist-packages/smbus2/smbus2.py", line 414, in write_byte
steering_emulator-dtp-1     |     ioctl(self.fd, I2C_SMBUS, msg)
steering_emulator-dtp-1     | OSError: [Errno 6] No such device or address
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | During handling of the above exception, another exception occurred:
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | Traceback (most recent call last):
steering_emulator-dtp-1     |   File "/root/catkin_ws/install/lib/picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py", line 191, in <module>
steering_emulator-dtp-1     |     clutchgear_emulator = AckermannClutchGearEmulator(
steering_emulator-dtp-1     |   File "/root/catkin_ws/install/lib/picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py", line 67, in __init__
steering_emulator-dtp-1     |     super(AckermannClutchGearEmulator, self).__init__(pwm_pin, i2c_port)
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/emulators/clutchgear.py", line 27, in __init__
steering_emulator-dtp-1     |     self.pwm_pin = {'channel': pwm_pin, 'i2c_port': i2c_port}
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/emulators/clutchgear.py", line 46, in pwm_pin
steering_emulator-dtp-1     |     self.__pwm_pin = PWM(
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 83, in __init__
steering_emulator-dtp-1     |     self.channel = channel
steering_emulator-dtp-1     |   File "/root/catkin_ws/src/core/picarx/src/picarx/pwm.py", line 125, in channel
steering_emulator-dtp-1     |     raise IOError("Can't send stuff.")
steering_emulator-dtp-1     | OSError: Can't send stuff.
shadow-dtp-1                | ... logging to /root/.ros/log/0924a26a-6ec0-11ef-80d4-0242ac120003/roslaunch-d8abc11ea095-1.log
shadow-dtp-1                | Checking log directory for disk usage. This may take a while.
shadow-dtp-1                | Press Ctrl-C to interrupt
shadow-dtp-1                | Done checking log file disk usage. Usage is <1GB.
shadow-dtp-1                |
ackermann_skill-dtp-1       | started roslaunch server http://76a6681eec60:36187/
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | SUMMARY
ackermann_skill-dtp-1       | ========
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | PARAMETERS
ackermann_skill-dtp-1       |  * /ackermann_skill/command_topic: /picarx/drive/com...
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_left_status_topic: /motor_left/status
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_left_topic: motor_left/speed
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_right_status_topic: /motor_right/status
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_right_topic: motor_right/speed
ackermann_skill-dtp-1       |  * /ackermann_skill/status_topic: drive/status
ackermann_skill-dtp-1       |  * /ackermann_skill/steering_status_topic: /steering/status
ackermann_skill-dtp-1       |  * /ackermann_skill/steering_topic: steering/angle
ackermann_skill-dtp-1       |  * /rosdistro: noetic
ackermann_skill-dtp-1       |  * /rosversion: 1.16.0
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | NODES
ackermann_skill-dtp-1       |   /
ackermann_skill-dtp-1       |     ackermann_skill (picarx_ackermann_drive/ackermann_drive_skill.py)
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | ROS_MASTER_URI=http://picarx:11311
ackermann_skill-dtp-1       |
picarx-gazebo-control-1     | ... logging to /root/.ros/log/09314c90-6ec0-11ef-84d9-0242ac120002/roslaunch-2127abb2723d-1.log
[…]
