### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
This error looks more familiar now.

On my windows system i2c-0 is the device I get after loading the module with modprobe.
On Ubuntu systems the i2c-0 could be there by default.

Try `I2C=/dev/i2c-1` in the simulation.env.

Since you have several i2c devices listed, you could try also the others if i2c-1 throws the same error.
