### 2024-09-09 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
**The comment I prepared yesterday after the commit but did not send:**

So, I tried to reproduce the error you posted and I am a bit clueless.

If the gpio pin is still active in the system, there error would be `OSError: [Errno 16] Device or resource busy`. If you use the wrong I2C device, the error would be `raise IOError("Can't send stuff.")`.

Did you activate the gpio-mockup module?

**END**

I will also add the error messages to the documentation.
