### 2024-09-09 c3c1367
added error messages to troubleshooting #4
