### 2024-09-10 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
> Try I2C=/dev/i2c-1 in the simulation.env.

> Since you have several i2c devices listed, you could try also the others if i2c-1 throws the same error.

I tried all 4 now. I keep getting this error:

<summary>Click to show log</summary>
<pre>
$ docker compose -f docker-compose-dtp.yml up
[+] Running 13/0
 ✔ Network picar-x_default                       Created                                           0.0s
 ✔ Container picar-x-picarx-1                    Created                                           0.0s
 ✔ Container picar-x-picarx-gazebo-1             Created                                           0.0s
 ✔ Container picar-x-motor_right-dtp-1           Created                                           0.0s
 ✔ Container picar-x-steering-dtp-1              Created                                           0.0s
 ✔ Container picar-x-motor_left-dtp-1            Created                                           0.0s
 ✔ Container picar-x-shadow-dtp-1                Created                                           0.0s
 ✔ Container picar-x-ackermann_skill-dtp-1       Crea...                                           0.0s
 ✔ Container picar-x-mqtt-dtp-1                  Created                                           0.0s
 ✔ Container picar-x-picarx-gazebo-control-1     Cr...                                             0.0s
 ✔ Container picar-x-motor_emulator_left-dtp-1   Created                                           0.0s
 ✔ Container picar-x-steering_emulator-dtp-1     Cr...                                             0.0s
 ✔ Container picar-x-motor_emulator_right-dtp-1  Created                                           0.0s
Attaching to ackermann_skill-dtp-1, motor_emulator_left-dtp-1, motor_emulator_right-dtp-1, motor_left-dtp-1, motor_right-dtp-1, mqtt-dtp-1, picarx-1, picarx-gazebo-1, picarx-gazebo-control-1, shadow-dtp-1, steering-dtp-1, steering_emulator-dtp-1
picarx-1                    | ... logging to /root/.ros/log/e3bb3c88-6f40-11ef-ab3a-0242ac120002/roslaunch-4dd5ea7aead3-1.log
picarx-1                    | Checking log directory for disk usage. This may take a while.
picarx-1                    | Press Ctrl-C to interrupt
picarx-1                    | Done checking log file disk usage. Usage is <1GB.
picarx-1                    |
picarx-gazebo-1             | ... logging to /root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/roslaunch-9df847c7cb1a-1.log
picarx-gazebo-1             | Checking log directory for disk usage. This may take a while.
picarx-gazebo-1             | Press Ctrl-C to interrupt
picarx-gazebo-1             | Done checking log file disk usage. Usage is <1GB.
picarx-gazebo-1             |
picarx-1                    | started roslaunch server http://picarx:45183/
picarx-1                    | ros_comm version 1.16.0
picarx-1                    |
picarx-1                    |
picarx-1                    | SUMMARY
picarx-1                    | ========
picarx-1                    |
picarx-1                    | PARAMETERS
picarx-1                    |  * /rosdistro: noetic
picarx-1                    |  * /rosversion: 1.16.0
picarx-1                    |
picarx-1                    | NODES
picarx-1                    |
picarx-1                    | auto-starting new master
picarx-1                    | process[master]: started with pid [32]
picarx-gazebo-1             | started roslaunch server http://picarx-gazebo:33363/
picarx-gazebo-1             |
picarx-gazebo-1             | SUMMARY
picarx-gazebo-1             | ========
picarx-gazebo-1             |
picarx-gazebo-1             | PARAMETERS
picarx-gazebo-1             |  * /gazebo/enable_ros_network: True
picarx-gazebo-1             |  * /robot_description: <?xml version="1....
picarx-gazebo-1             |  * /rosdistro: noetic
picarx-gazebo-1             |  * /rosversion: 1.16.0
picarx-gazebo-1             |  * /use_sim_time: True
picarx-gazebo-1             |
picarx-gazebo-1             | NODES
picarx-gazebo-1             |   /
picarx-gazebo-1             |     gazebo (gazebo_ros/gzserver)
picarx-gazebo-1             |     gazebo_gui (gazebo_ros/gzclient)
picarx-gazebo-1             |     spawn_urdf (gazebo_ros/spawn_model)
picarx-gazebo-1             |
picarx-gazebo-1             | auto-starting new master
picarx-gazebo-1             | process[master]: started with pid [34]
picarx-1                    | ROS_MASTER_URI=http://picarx:11311/
picarx-1                    |
picarx-1                    | setting /run_id to e3bb3c88-6f40-11ef-ab3a-0242ac120002
picarx-1                    | process[rosout-1]: started with pid [42]
picarx-1                    | started core service [/rosout]
picarx-gazebo-1             | ROS_MASTER_URI=http://picarx-gazebo:11312
picarx-gazebo-1             |
picarx-gazebo-1             | setting /run_id to e3bdd0ec-6f40-11ef-b42d-0242ac120003
picarx-gazebo-1             | process[rosout-1]: started with pid [44]
picarx-gazebo-1             | started core service [/rosout]
picarx-gazebo-1             | process[gazebo-2]: started with pid [47]
picarx-gazebo-1             | process[gazebo_gui-3]: started with pid [54]
picarx-gazebo-1             | process[spawn_urdf-4]: started with pid [61]
picarx-gazebo-1             | [ INFO] [1725950996.959012178]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1725950996.959783662]: waitForService: Service [/gazebo/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1725950997.022607587]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1725950997.023435489]: waitForService: Service [/gazebo_gui/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1725950997.103391208]: waitForService: Service [/gazebo/set_physics_properties] is now available.
picarx-gazebo-1             | [ INFO] [1725950997.115105877, 0.008000000]: Physics dynamic reconfigure ready.
picarx-gazebo-1             | Aborted (core dumped)
picarx-gazebo-1             | [gazebo_gui-3] process has died [pid 54, exit code 134, cmd /opt/ros/noetic/lib/gazebo_ros/gzclient __name:=gazebo_gui __log:=/root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/gazebo_gui-3.log].
picarx-gazebo-1             | log file: /root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/gazebo_gui-3*.log
picarx-gazebo-1             | [ INFO] [1725950997.403845635, 0.177000000]: Loading gazebo_ros_control plugin
picarx-gazebo-1             | [ INFO] [1725950997.403911420, 0.177000000]: Starting gazebo_ros_control plugin in namespace: /picarx
picarx-gazebo-1             | [ INFO] [1725950997.404304573, 0.177000000]: gazebo_ros_control plugin is waiting for model URDF in parameter [/robot_description] on the ROS param server.
picarx-gazebo-1             | [ERROR] [1725950997.546985720, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_right
picarx-gazebo-1             | [ERROR] [1725950997.548177256, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_left
picarx-gazebo-1             | [ERROR] [1725950997.549241592, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_left
picarx-gazebo-1             | [ERROR] [1725950997.550161392, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_right
picarx-gazebo-1             | [ INFO] [1725950997.553918379, 0.177000000]: Loaded gazebo_ros_control.
picarx-gazebo-1             | [spawn_urdf-4] process has finished cleanly
picarx-gazebo-1             | log file: /root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/spawn_urdf-4*.log
picarx-gazebo-control-1     | ... logging to /root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/roslaunch-8a5dc8c92108-1.log
picarx-gazebo-control-1     | Checking log directory for disk usage. This may take a while.
picarx-gazebo-control-1     | Press Ctrl-C to interrupt
picarx-gazebo-control-1     | Done checking log file disk usage. Usage is <1GB.
picarx-gazebo-control-1     |
shadow-dtp-1                | ... logging to /root/.ros/log/e3bb3c88-6f40-11ef-ab3a-0242ac120002/roslaunch-bbfbd3b38ba1-1.log
shadow-dtp-1                | Checking log directory for disk usage. This may take a while.
shadow-dtp-1                | Press Ctrl-C to interrupt
shadow-dtp-1                | Done checking log file disk usage. Usage is <1GB.
shadow-dtp-1                |
picarx-gazebo-control-1     | started roslaunch server http://8a5dc8c92108:44997/
picarx-gazebo-control-1     |
picarx-gazebo-control-1     | SUMMARY
picarx-gazebo-control-1     | ========
picarx-gazebo-control-1     |
picarx-gazebo-control-1     | PARAMETERS
picarx-gazebo-control-1     |  * /picarx/joint_state_controller/publish_rate: 50
picarx-gazebo-control-1     |  * /picarx/joint_state_controller/type: joint_state_contr...
picarx-gazebo-control-1     |  * /picarx/rear_left_wheel_controller/joint: joint_wheel_left
picarx-gazebo-control-1     |  * /picarx/rear_left_wheel_controller/pid/d: 10.0
picarx-gazebo-control-1     |  * /picarx/rear_left_wheel_controller/pid/i: 0.01
picarx-gazebo-control-1     |  * /picarx/rear_left_wheel_controller/pid/p: 100.0
picarx-gazebo-control-1     |  * /picarx/rear_left_wheel_controller/type: velocity_controll...
picarx-gazebo-control-1     |  * /picarx/rear_right_wheel_controller/joint: joint_wheel_right
picarx-gazebo-control-1     |  * /picarx/rear_right_wheel_controller/pid/d: 10.0
picarx-gazebo-control-1     |  * /picarx/rear_right_wheel_controller/pid/i: 0.01
picarx-gazebo-control-1     |  * /picarx/rear_right_wheel_controller/pid/p: 100.0
picarx-gazebo-control-1     |  * /picarx/rear_right_wheel_controller/type: velocity_controll...
picarx-gazebo-control-1     |  * /picarx/steer_left_controller/joint: steer_left
picarx-gazebo-control-1     |  * /picarx/steer_left_controller/type: position_controll...
picarx-gazebo-control-1     |  * /picarx/steer_right_controller/joint: steer_right
picarx-gazebo-control-1     |  * /picarx/steer_right_controller/type: position_controll...
picarx-gazebo-control-1     |  * /rosdistro: noetic
picarx-gazebo-control-1     |  * /rosversion: 1.16.0
picarx-gazebo-control-1     |
picarx-gazebo-control-1     | NODES
picarx-gazebo-control-1     |   /
picarx-gazebo-control-1     |     robot_state_publisher (robot_state_publisher/robot_state_publisher)
picarx-gazebo-control-1     |   /picarx/
picarx-gazebo-control-1     |     controller_spawner (controller_manager/spawner)
picarx-gazebo-control-1     |
picarx-gazebo-control-1     | ROS_MASTER_URI=http://picarx-gazebo:11312
picarx-gazebo-control-1     |
steering-dtp-1              | ... logging to /root/.ros/log/e3bb3c88-6f40-11ef-ab3a-0242ac120002/roslaunch-ea81f50b53df-1.log
steering-dtp-1              | Checking log directory for disk usage. This may take a while.
motor_left-dtp-1            | ... logging to /root/.ros/log/e3bb3c88-6f40-11ef-ab3a-0242ac120002/roslaunch-67ed9c753c9c-1.log
steering-dtp-1              | Press Ctrl-C to interrupt
motor_left-dtp-1            | Checking log directory for disk usage. This may take a while.
motor_left-dtp-1            | Press Ctrl-C to interrupt
motor_left-dtp-1            | Done checking log file disk usage. Usage is <1GB.
steering-dtp-1              | Done checking log file disk usage. Usage is <1GB.
motor_left-dtp-1            |
steering-dtp-1              |
steering_emulator-dtp-1     | ... logging to /root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/roslaunch-c4146a4e4e2f-1.log
steering_emulator-dtp-1     | Checking log directory for disk usage. This may take a while.
steering_emulator-dtp-1     | Press Ctrl-C to interrupt
steering_emulator-dtp-1     | Done checking log file disk usage. Usage is <1GB.
steering_emulator-dtp-1     |
mqtt-dtp-1                  | ... logging to /root/.ros/log/e3bb3c88-6f40-11ef-ab3a-0242ac120002/roslaunch-02c9a4524a43-1.log
mqtt-dtp-1                  | Checking log directory for disk usage. This may take a while.
mqtt-dtp-1                  | Press Ctrl-C to interrupt
mqtt-dtp-1                  | Done checking log file disk usage. Usage is <1GB.
mqtt-dtp-1                  |
shadow-dtp-1                | started roslaunch server http://bbfbd3b38ba1:41793/
shadow-dtp-1                |
shadow-dtp-1                | SUMMARY
shadow-dtp-1                | ========
shadow-dtp-1                |
shadow-dtp-1                | PARAMETERS
shadow-dtp-1                |  * /rosdistro: noetic
shadow-dtp-1                |  * /rosversion: 1.16.0
shadow-dtp-1                |
shadow-dtp-1                | NODES
shadow-dtp-1                |   /
shadow-dtp-1                |     PTControlDistributor (arches_core/digitalshadow_control_pt.py)
shadow-dtp-1                |     PTDataCollector (arches_core/digitalshadow_data_pt.py)
shadow-dtp-1                |
shadow-dtp-1                | ROS_MASTER_URI=http://picarx:11311
shadow-dtp-1                |
picarx-gazebo-control-1     | process[picarx/controller_spawner-1]: started with pid [32]
picarx-gazebo-control-1     | process[robot_state_publisher-2]: started with pid [33]
shadow-dtp-1                | process[PTDataCollector-1]: started with pid [32]
shadow-dtp-1                | process[PTControlDistributor-2]: started with pid [33]
steering-dtp-1              | started roslaunch server http://ea81f50b53df:39149/
steering-dtp-1              |
steering-dtp-1              | SUMMARY
steering-dtp-1              | ========
steering-dtp-1              |
steering-dtp-1              | PARAMETERS
steering-dtp-1              |  * /ackermann_clutchgear_driver/i2c_port: /dev/i2c-1
steering-dtp-1              |  * /ackermann_clutchgear_driver/name: ackermann_clutchg...
steering-dtp-1              |  * /ackermann_clutchgear_driver/pwm_pin: P2
steering-dtp-1              |  * /ackermann_clutchgear_driver/steering_status_topic: steering/status
steering-dtp-1              |  * /ackermann_clutchgear_driver/steering_topic: steering/angle
steering-dtp-1              |  * /rosdistro: noetic
steering-dtp-1              |  * /rosversion: 1.16.0
steering-dtp-1              |
steering-dtp-1              | NODES
steering-dtp-1              |   /
steering-dtp-1              |     ackermann_clutchgear_driver (picarx_clutchgear_driver/ackermann_clutchgear.py)
steering-dtp-1              |
steering-dtp-1              | ROS_MASTER_URI=http://picarx:11311
steering-dtp-1              |
motor_left-dtp-1            | started roslaunch server http://67ed9c753c9c:36499/
motor_emulator_left-dtp-1   | ... logging to /root/.ros/log/e3bdd0ec-6f40-11ef-b42d-0242ac120003/roslaunch-f2c3d703edea-1.log
motor_emulator_left-dtp-1   | Checking log directory for disk usage. This may take a while.
motor_emulator_left-dtp-1   | Press Ctrl-C to interrupt
motor_left-dtp-1            |
motor_left-dtp-1            | SUMMARY
motor_left-dtp-1            | ========
motor_left-dtp-1            |
motor_left-dtp-1            | PARAMETERS
motor_left-dtp-1            |  * /motor_left/command_topic: motor_left/speed
motor_left-dtp-1            |  * /motor_left/direction_pin: 24
motor_left-dtp-1            |  * /motor_left/i2c_port: /dev/i2c-1
motor_left-dtp-1            |  * /motor_left/motor_side: 0
motor_left-dtp-1            |  * /motor_left/name: motor_left
motor_left-dtp-1            |  * /motor_left/pwm_pin: P12
motor_left-dtp-1            |  * /motor_left/status_topic: motor_left/status
motor_left-dtp-1            |  * /rosdistro: noetic
motor_left-dtp-1            |  * /rosversion: 1.16.0
motor_left-dtp-1            |
motor_left-dtp-1            | NODES
motor_left-dtp-1            |   /
motor_left-dtp-1            |     motor_left (picarx_dcmotor_driver/dcmotor_driver.py)
motor_left-dtp-1            |
motor_emulator_left-dtp-1   | Done checking log file disk usage. Usage is <1GB.
motor_emulator_left-dtp-1   |
motor_left-dtp-1            | ROS_MASTER_URI=http://picarx:11311
motor_left-dtp-1            |
steering_emulator-dtp-1     | started roslaunch server http://c4146a4e4e2f:33883/
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | SUMMARY
steering_emulator-dtp-1     | ========
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | PARAMETERS
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/i2c_port: /dev/i2c-0
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/left_steer_topic: /picarx/steer_lef...
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/name: ackermann_clutchg...
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/pwm_pin: P2
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/right_steer_topic: /picarx/steer_rig...
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/wheel_base: 13.5
steering_emulator-dtp-1     |  * /ackermann_clutchgear_emulator/wheel_track: 9.2
steering_emulator-dtp-1     |  * /rosdistro: noetic
steering_emulator-dtp-1     |  * /rosversion: 1.16.0
steering_emulator-dtp-1     |
steering_emulator-dtp-1     | NODES
steering_emulator-dtp-1     |   /
steering_emulator-dtp-1     |     ackermann_clutchgear_emulator (picarx_clutchgear_emulator/ackermann_clutchgear_emulator.py)
steering_emulator-dtp-1     |
steering-dtp-1              | process[ackermann_clutchgear_driver-1]: started with pid [32]
steering_emulator-dtp-1     | ROS_MASTER_URI=http://picarx-gazebo:11312
steering_emulator-dtp-1     |
motor_left-dtp-1            | process[motor_left-1]: started with pid [32]
steering_emulator-dtp-1     | process[ackermann_clutchgear_emulator-1]: started with pid [32]
mqtt-dtp-1                  | started roslaunch server http://02c9a4524a43:43017/
mqtt-dtp-1                  |
mqtt-dtp-1                  | SUMMARY
mqtt-dtp-1                  | ========
mqtt-dtp-1                  |
mqtt-dtp-1                  | PARAMETERS
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/client/clean_session: True
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/client/client_id:
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/client/protocol: MQTTv311
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/client/transport: tcp
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/client/userdata: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/connect/bind_address:
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/connect/host: mqtt-server
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/connect/keepalive: 60
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/connect/port: 1883
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/credentials/password: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/credentials/username: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/tls/ca_certs: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/tls/cert_reqs: True
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/tls/certfile: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/tls/ciphers: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/tls/keyfile: None
mqtt-dtp-1                  |  * /MQTTTwinBridge/config/tls/tls_version: True
mqtt-dtp-1                  |  * /rosdistro: noetic
mqtt-dtp-1                  |  * /rosversion: 1.16.0
mqtt-dtp-1                  |
mqtt-dtp-1                  | NODES
mqtt-dtp-1                  |   /
mqtt-dtp-1                  |     MQTTTwinBridge (arches_mqtt_bridge/twin_bridge_pt.py)
mqtt-dtp-1                  |
mqtt-dtp-1                  | ROS_MASTER_URI=http://picarx:11311
mqtt-dtp-1                  |
mqtt-dtp-1                  | process[MQTTTwinBridge-1]: started with pid [32]
motor_emulator_left-dtp-1   | started roslaunch server http://f2c3d703edea:45917/
motor_emulator_left-dtp-1   |
motor_emulator_left-dtp-1   | SUMMARY
motor_emulator_left-dtp-1   | ========
motor_emulator_left-dtp-1   |
motor_emulator_left-dtp-1   | PARAMETERS
[…]
