### 2024-09-10 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
There is a way to check the registers on the I2C.

The general way is:
1. install with `sudo apt install i2c-tools`
2. Execute `i2cdetect -l`
3. Execute `i2cdetect -l -r 0` to get the registers for `/dev/i2c-0`. Replace the 0 with the number of the other devices.

I do not have access to my system right now, but the result should be something like here: [using_i2c](http://www.merkles.com/Using_I2C/1_i2cdetect.html). Only one register should be active (different from --).

I further saw, that the -dtp.yml also uses the picarx.env. There you should adjust the I2C value, too. The value should match the simulation.env. But I do not think that this will solve the current problem.

Maybe there is an docker error with default environment variable values, try to execute docker compose with:

```console
I2C=/dev/i2c-1 docker compose -f docker-compose-dtp.yml up
```

And again, replace the -1 with any other device number also set in the *.env.

Your question regarding the activation after reboot: it is possible to place the stuff in autostart, but did not try this. On the RPi the modules are activated by default.

**EDIT**: There is also the file [setup-gpio-i2c-emulation.bash](https://github.com/cau-se/ARCHES-PiCar-X/blob/main/PiCar-X/setup-gpio-i2c-emulation.bash), where all three commands are executed. However, my experience is, that sometimes these commands are not executed properly.
