### 2024-09-10 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
```
$ i2cdetect -l
i2c-3	unknown   	SMBus stub driver               	N/A
i2c-1	unknown   	Synopsys DesignWare I2C adapter 	N/A
i2c-2	unknown   	Synopsys DesignWare I2C adapter 	N/A
i2c-0	unknown   	SMBus I801 adapter at efa0      	N/A

$ sudo i2cdetect -r 0
WARNING! This program can confuse your I2C bus, cause data loss and worse!
I will probe file /dev/i2c-0 using receive byte commands.
I will probe address range 0x03-0x77.
Continue? [Y/n] Y
     0  1  2  3  4  5  6  7  8  9  a  b  c  d  e  f
00:          -- -- -- -- -- -- -- -- -- -- -- -- --
10: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
20: 20 -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
30: 30 -- -- -- -- -- UU UU -- -- -- -- -- -- -- --
40: -- -- -- -- 44 -- -- -- -- -- -- -- -- -- -- --
50: -- UU -- UU -- -- -- -- -- -- -- -- -- -- -- --
60: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
70: -- -- -- -- -- -- -- --

$ sudo i2cdetect -r 1
WARNING! This program can confuse your I2C bus, cause data loss and worse!
I will probe file /dev/i2c-1 using receive byte commands.
I will probe address range 0x03-0x77.
Continue? [Y/n] Y
     0  1  2  3  4  5  6  7  8  9  a  b  c  d  e  f
00:          -- -- -- -- -- -- -- -- -- -- -- -- --
10: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
20: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
30: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
40: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
50: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
60: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
70: -- -- -- -- -- -- -- --

$ sudo i2cdetect -r 2
WARNING! This program can confuse your I2C bus, cause data loss and worse!
I will probe file /dev/i2c-2 using receive byte commands.
I will probe address range 0x03-0x77.
Continue? [Y/n] Y
     0  1  2  3  4  5  6  7  8  9  a  b  c  d  e  f
00:          -- -- -- -- -- -- -- -- -- -- -- -- --
10: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
20: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
30: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
40: 40 -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
50: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
60: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
70: -- -- -- -- -- -- -- --

$ sudo i2cdetect -r 3
WARNING! This program can confuse your I2C bus, cause data loss and worse!
I will probe file /dev/i2c-3 using receive byte commands.
I will probe address range 0x03-0x77.
Continue? [Y/n] Y
     0  1  2  3  4  5  6  7  8  9  a  b  c  d  e  f
00:          -- -- -- -- -- -- -- -- -- -- -- -- --
10: -- -- -- -- 14 -- -- -- -- -- -- -- -- -- -- --
20: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
30: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
40: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
50: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
60: -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- --
70: -- -- -- -- -- -- -- --
```

So I tried 2 and 3. Changed simulation.env and picar.env and started with `I2C=/dev/i2c-X docker compose -f docker-compose-dtp.yml up`. 2 failed again with the same error. 3 works kind of:

<summary>Click to show the full log.</summary>
<pre>
$ I2C=/dev/i2c-3 docker compose -f docker-compose-dtp.yml up
[+] Running 13/0
 ✔ Network picar-x_default                       Created                                         0.0s
 ✔ Container picar-x-picarx-gazebo-1             Created                                         0.0s
 ✔ Container picar-x-picarx-1                    Created                                         0.0s
 ✔ Container picar-x-ackermann_skill-dtp-1       Cr...                                           0.0s
 ✔ Container picar-x-mqtt-dtp-1                  Created                                         0.0s
 ✔ Container picar-x-shadow-dtp-1                Created                                         0.0s
 ✔ Container picar-x-motor_left-dtp-1            Created                                         0.0s
 ✔ Container picar-x-motor_right-dtp-1           Create...                                       0.0s
 ✔ Container picar-x-steering-dtp-1              Created                                         0.0s
 ✔ Container picar-x-steering_emulator-dtp-1     Created                                         0.0s
 ✔ Container picar-x-motor_emulator_left-dtp-1   Created                                         0.0s
 ✔ Container picar-x-motor_emulator_right-dtp-1  Created                                         0.0s
 ✔ Container picar-x-picarx-gazebo-control-1     Created                                         0.0s
Attaching to ackermann_skill-dtp-1, motor_emulator_left-dtp-1, motor_emulator_right-dtp-1, motor_left-dtp-1, motor_right-dtp-1, mqtt-dtp-1, picarx-1, picarx-gazebo-1, picarx-gazebo-control-1, shadow-dtp-1, steering-dtp-1, steering_emulator-dtp-1
picarx-1                    | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-d9ae73d76881-1.log
picarx-1                    | Checking log directory for disk usage. This may take a while.
picarx-1                    | Press Ctrl-C to interrupt
picarx-1                    | Done checking log file disk usage. Usage is <1GB.
picarx-1                    |
picarx-gazebo-1             | ... logging to /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/roslaunch-bbc2e7b99351-1.log
picarx-gazebo-1             | Checking log directory for disk usage. This may take a while.
picarx-gazebo-1             | Press Ctrl-C to interrupt
picarx-gazebo-1             | Done checking log file disk usage. Usage is <1GB.
picarx-gazebo-1             |
picarx-1                    | started roslaunch server http://picarx:35891/
picarx-1                    | ros_comm version 1.16.0
picarx-1                    |
picarx-1                    |
picarx-1                    | SUMMARY
picarx-1                    | ========
picarx-1                    |
picarx-1                    | PARAMETERS
picarx-1                    |  * /rosdistro: noetic
picarx-1                    |  * /rosversion: 1.16.0
picarx-1                    |
picarx-1                    | NODES
picarx-1                    |
picarx-1                    | auto-starting new master
picarx-1                    | process[master]: started with pid [32]
picarx-gazebo-1             | started roslaunch server http://picarx-gazebo:43127/
picarx-gazebo-1             |
picarx-gazebo-1             | SUMMARY
picarx-gazebo-1             | ========
picarx-gazebo-1             |
picarx-gazebo-1             | PARAMETERS
picarx-gazebo-1             |  * /gazebo/enable_ros_network: True
picarx-gazebo-1             |  * /robot_description: <?xml version="1....
picarx-gazebo-1             |  * /rosdistro: noetic
picarx-gazebo-1             |  * /rosversion: 1.16.0
picarx-gazebo-1             |  * /use_sim_time: True
picarx-gazebo-1             |
picarx-gazebo-1             | NODES
picarx-gazebo-1             |   /
picarx-gazebo-1             |     gazebo (gazebo_ros/gzserver)
picarx-gazebo-1             |     gazebo_gui (gazebo_ros/gzclient)
picarx-gazebo-1             |     spawn_urdf (gazebo_ros/spawn_model)
picarx-gazebo-1             |
picarx-gazebo-1             | auto-starting new master
picarx-gazebo-1             | process[master]: started with pid [34]
picarx-1                    | ROS_MASTER_URI=http://picarx:11311/
picarx-1                    |
picarx-1                    | setting /run_id to ccce3aac-6f54-11ef-b7c4-0242ac120002
picarx-1                    | process[rosout-1]: started with pid [42]
picarx-1                    | started core service [/rosout]
picarx-gazebo-1             | ROS_MASTER_URI=http://picarx-gazebo:11312
picarx-gazebo-1             |
picarx-gazebo-1             | setting /run_id to ccd0ffbc-6f54-11ef-bae6-0242ac120003
picarx-gazebo-1             | process[rosout-1]: started with pid [44]
picarx-gazebo-1             | started core service [/rosout]
picarx-gazebo-1             | process[gazebo-2]: started with pid [47]
picarx-gazebo-1             | process[gazebo_gui-3]: started with pid [55]
picarx-gazebo-1             | process[spawn_urdf-4]: started with pid [61]
picarx-gazebo-1             | [ INFO] [1725959548.435797218]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1725959548.436551118]: waitForService: Service [/gazebo/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1725959548.493118764]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1725959548.493930672]: waitForService: Service [/gazebo_gui/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1725959548.600178142, 0.005000000]: waitForService: Service [/gazebo/set_physics_properties] is now available.
picarx-gazebo-1             | [ INFO] [1725959548.612152766, 0.017000000]: Physics dynamic reconfigure ready.
picarx-gazebo-1             | Aborted (core dumped)
picarx-gazebo-1             | [gazebo_gui-3] process has died [pid 55, exit code 134, cmd /opt/ros/noetic/lib/gazebo_ros/gzclient __name:=gazebo_gui __log:=/root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/gazebo_gui-3.log].
picarx-gazebo-1             | log file: /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/gazebo_gui-3*.log
picarx-gazebo-1             | [ INFO] [1725959548.891192520, 0.177000000]: Loading gazebo_ros_control plugin
picarx-gazebo-1             | [ INFO] [1725959548.891259314, 0.177000000]: Starting gazebo_ros_control plugin in namespace: /picarx
picarx-gazebo-1             | [ INFO] [1725959548.891684227, 0.177000000]: gazebo_ros_control plugin is waiting for model URDF in parameter [/robot_description] on the ROS param server.
picarx-gazebo-1             | [ERROR] [1725959549.032439113, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_right
picarx-gazebo-1             | [ERROR] [1725959549.033608974, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_left
picarx-gazebo-1             | [ERROR] [1725959549.034764152, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_left
picarx-gazebo-1             | [ERROR] [1725959549.035633302, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_right
picarx-gazebo-1             | [ INFO] [1725959549.038213899, 0.177000000]: Loaded gazebo_ros_control.
picarx-gazebo-1             | [spawn_urdf-4] process has finished cleanly
picarx-gazebo-1             | log file: /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/spawn_urdf-4*.log
motor_emulator_right-dtp-1  | ... logging to /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/roslaunch-749298c1bf3a-1.log
motor_emulator_right-dtp-1  | Checking log directory for disk usage. This may take a while.
motor_emulator_right-dtp-1  | Press Ctrl-C to interrupt
motor_emulator_right-dtp-1  | Done checking log file disk usage. Usage is <1GB.
motor_emulator_right-dtp-1  |
motor_left-dtp-1            | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-0c2b542f91ea-1.log
motor_left-dtp-1            | Checking log directory for disk usage. This may take a while.
motor_left-dtp-1            | Press Ctrl-C to interrupt
motor_left-dtp-1            | Done checking log file disk usage. Usage is <1GB.
motor_left-dtp-1            |
motor_right-dtp-1           | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-516e49f6338d-1.log
motor_right-dtp-1           | Checking log directory for disk usage. This may take a while.
motor_right-dtp-1           | Press Ctrl-C to interrupt
motor_right-dtp-1           | Done checking log file disk usage. Usage is <1GB.
motor_right-dtp-1           |
ackermann_skill-dtp-1       | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-9475d867f790-1.log
ackermann_skill-dtp-1       | Checking log directory for disk usage. This may take a while.
ackermann_skill-dtp-1       | Press Ctrl-C to interrupt
ackermann_skill-dtp-1       | Done checking log file disk usage. Usage is <1GB.
ackermann_skill-dtp-1       |
motor_emulator_right-dtp-1  | started roslaunch server http://749298c1bf3a:46331/
motor_emulator_right-dtp-1  |
motor_emulator_right-dtp-1  | SUMMARY
motor_emulator_right-dtp-1  | ========
motor_emulator_right-dtp-1  |
motor_emulator_right-dtp-1  | PARAMETERS
motor_emulator_right-dtp-1  |  * /motor_emulator_right/controller_publisher_topic: /picarx/rear_righ...
motor_emulator_right-dtp-1  |  * /motor_emulator_right/direction_pin: 23
motor_emulator_right-dtp-1  |  * /motor_emulator_right/i2c_port: /dev/i2c-3
motor_emulator_right-dtp-1  |  * /motor_emulator_right/motor_side: 1
motor_emulator_right-dtp-1  |  * /motor_emulator_right/name: motor_emulator_right
motor_emulator_right-dtp-1  |  * /motor_emulator_right/pwm_pin: P13
motor_emulator_right-dtp-1  |  * /rosdistro: noetic
motor_emulator_right-dtp-1  |  * /rosversion: 1.16.0
motor_emulator_right-dtp-1  |
motor_emulator_right-dtp-1  | NODES
motor_emulator_right-dtp-1  |   /
motor_emulator_right-dtp-1  |     motor_emulator_right (picarx_dcmotor_emulator/dcmotor_emulator.py)
motor_emulator_right-dtp-1  |
motor_left-dtp-1            | started roslaunch server http://0c2b542f91ea:46159/
motor_left-dtp-1            |
motor_left-dtp-1            | SUMMARY
motor_left-dtp-1            | ========
motor_left-dtp-1            |
motor_left-dtp-1            | PARAMETERS
motor_left-dtp-1            |  * /motor_left/command_topic: motor_left/speed
motor_left-dtp-1            |  * /motor_left/direction_pin: 24
motor_left-dtp-1            |  * /motor_left/i2c_port: /dev/i2c-3
motor_left-dtp-1            |  * /motor_left/motor_side: 0
motor_left-dtp-1            |  * /motor_left/name: motor_left
motor_left-dtp-1            |  * /motor_left/pwm_pin: P12
motor_left-dtp-1            |  * /motor_left/status_topic: motor_left/status
motor_left-dtp-1            |  * /rosdistro: noetic
motor_left-dtp-1            |  * /rosversion: 1.16.0
motor_left-dtp-1            |
motor_left-dtp-1            | NODES
motor_left-dtp-1            |   /
motor_left-dtp-1            |     motor_left (picarx_dcmotor_driver/dcmotor_driver.py)
motor_left-dtp-1            |
motor_right-dtp-1           | started roslaunch server http://516e49f6338d:36785/
motor_right-dtp-1           |
motor_right-dtp-1           | SUMMARY
motor_right-dtp-1           | ========
motor_right-dtp-1           |
motor_right-dtp-1           | PARAMETERS
motor_right-dtp-1           |  * /motor_right/command_topic: motor_right/speed
motor_right-dtp-1           |  * /motor_right/direction_pin: 23
motor_right-dtp-1           |  * /motor_right/i2c_port: /dev/i2c-3
motor_right-dtp-1           |  * /motor_right/motor_side: 1
motor_right-dtp-1           |  * /motor_right/name: motor_right
motor_right-dtp-1           |  * /motor_right/pwm_pin: P13
motor_right-dtp-1           |  * /motor_right/status_topic: motor_right/status
motor_right-dtp-1           |  * /rosdistro: noetic
motor_right-dtp-1           |  * /rosversion: 1.16.0
motor_right-dtp-1           |
motor_right-dtp-1           | NODES
motor_right-dtp-1           |   /
motor_right-dtp-1           |     motor_right (picarx_dcmotor_driver/dcmotor_driver.py)
motor_right-dtp-1           |
motor_emulator_right-dtp-1  | ROS_MASTER_URI=http://picarx-gazebo:11312
motor_emulator_right-dtp-1  |
ackermann_skill-dtp-1       | started roslaunch server http://9475d867f790:42783/
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | SUMMARY
ackermann_skill-dtp-1       | ========
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | PARAMETERS
ackermann_skill-dtp-1       |  * /ackermann_skill/command_topic: /picarx/drive/com...
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_left_status_topic: /motor_left/status
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_left_topic: motor_left/speed
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_right_status_topic: /motor_right/status
ackermann_skill-dtp-1       |  * /ackermann_skill/motor_right_topic: motor_right/speed
ackermann_skill-dtp-1       |  * /ackermann_skill/status_topic: drive/status
ackermann_skill-dtp-1       |  * /ackermann_skill/steering_status_topic: /steering/status
ackermann_skill-dtp-1       |  * /ackermann_skill/steering_topic: steering/angle
ackermann_skill-dtp-1       |  * /rosdistro: noetic
ackermann_skill-dtp-1       |  * /rosversion: 1.16.0
ackermann_skill-dtp-1       |
ackermann_skill-dtp-1       | NODES
ackermann_skill-dtp-1       |   /
ackermann_skill-dtp-1       |     ackermann_skill (picarx_ackermann_drive/ackermann_drive_skill.py)
ackermann_skill-dtp-1       |
motor_right-dtp-1           | ROS_MASTER_URI=http://picarx:11311
motor_right-dtp-1           |
ackermann_skill-dtp-1       | ROS_MASTER_URI=http://picarx:11311
ackermann_skill-dtp-1       |
motor_left-dtp-1            | ROS_MASTER_URI=http://picarx:11311
motor_left-dtp-1            |
motor_emulator_right-dtp-1  | process[motor_emulator_right-1]: started with pid [31]
motor_left-dtp-1            | process[motor_left-1]: started with pid [32]
motor_right-dtp-1           | process[motor_right-1]: started with pid [31]
ackermann_skill-dtp-1       | process[ackermann_skill-1]: started with pid [32]
steering_emulator-dtp-1     | ... logging to /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/roslaunch-3d59df6199be-1.log
steering_emulator-dtp-1     | Checking log directory for disk usage. This may take a while.
steering_emulator-dtp-1     | Press Ctrl-C to interrupt
steering_emulator-dtp-1     | Done checking log file disk usage. Usage is <1GB.
steering_emulator-dtp-1     |
steering-dtp-1              | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-695c973d5738-1.log
steering-dtp-1              | Checking log directory for disk usage. This may take a while.
steering-dtp-1              | Press Ctrl-C to interrupt
steering-dtp-1              | Done checking log file disk usage. Usage is <1GB.
steering-dtp-1              |
shadow-dtp-1                | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-6e7598115ecb-1.log
shadow-dtp-1                | Checking log directory for disk usage. This may take a while.
shadow-dtp-1                | Press Ctrl-C to interrupt
shadow-dtp-1                | Done checking log file disk usage. Usage is <1GB.
shadow-dtp-1                |
picarx-gazebo-control-1     | ... logging to /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/roslaunch-1f066e8180a9-1.log
picarx-gazebo-control-1     | Checking log directory for disk usage. This may take a while.
picarx-gazebo-control-1     | Press Ctrl-C to interrupt
picarx-gazebo-control-1     | Done checking log file disk usage. Usage is <1GB.
picarx-gazebo-control-1     |
motor_emulator_left-dtp-1   | ... logging to /root/.ros/log/ccd0ffbc-6f54-11ef-bae6-0242ac120003/roslaunch-751f8b484dbf-1.log
motor_emulator_left-dtp-1   | Checking log directory for disk usage. This may take a while.
motor_emulator_left-dtp-1   | Press Ctrl-C to interrupt
motor_emulator_left-dtp-1   | Done checking log file disk usage. Usage is <1GB.
motor_emulator_left-dtp-1   |
mqtt-dtp-1                  | ... logging to /root/.ros/log/ccce3aac-6f54-11ef-b7c4-0242ac120002/roslaunch-85785f875c7b-1.log
mqtt-dtp-1                  | Checking log directory for disk usage. This may take a while.
mqtt-dtp-1                  | Press Ctrl-C to interrupt
mqtt-dtp-1                  | Done checking log file disk usage. Usage is <1GB.
mqtt-dtp-1                  |
steering-dtp-1              | started roslaunch server http://695c973d5738:32801/
steering-dtp-1              |
steering-dtp-1              | SUMMARY
steering-dtp-1              | ========
steering-dtp-1              |
steering-dtp-1              | PARAMETERS
steering-dtp-1              |  * /ackermann_clutchgear_driver/i2c_port: /dev/i2c-3
[…]
