### 2024-09-10 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
Additional remarks:
* it's a bit annoying to copy the commands from "Getting started" because they are all indented by 4 spaces
* I don't have any visual confirmation after sending commands to the picar, is there a guide to set up the Gazebo GUI?
