### 2024-09-10 fe3dbbd
removed whitespaces in docu and added xhost activation for gazebo gui #4
