+++ docs/digitaltwins/ptds.md
@@ -15,58 +15,58 @@ In this example, we demonstrate how the PiCar-X can be connected to a digital sh
 We assume at this point that the digital shadow runs on a X64 Linux system (e.g. a server).
 
 ```console
-    # Build and execute the Docker Containers
-    docker compose -f docker-compose-core.yml build 
-    docker compose -f docker-compose-ds.yml build 
-    docker compose -f docker-compose-ds.yml up
+# Build and execute the Docker Containers
+docker compose -f docker-compose-core.yml build 
+docker compose -f docker-compose-ds.yml build 
+docker compose -f docker-compose-ds.yml up
 ```
 
 
 ## Start the Phyiscal Twin
 <div class="tab-container" id="activaterpi">
   <ul class="tab-list">
-    <li class="tab active" data-tab="tab3-1">arm32v7 (RPi3)</li>
-    <li class="tab" data-tab="tab3-2">arm64v8 (RPi4)</li>
+<li class="tab active" data-tab="tab3-1">arm32v7 (RPi3)</li>
+<li class="tab" data-tab="tab3-2">arm64v8 (RPi4)</li>
   </ul>
   <div class="tab-content active" id="tab3-1">
   {% highlight console %}
-    # First copy all content to the RaspberryPi4
-    scp -r ./ <user>@<picarx-ip>:~/
+# First copy all content to the RaspberryPi4
+scp -r ./ <user>@<picarx-ip>:~/
 
-    # Login to the RaspberryPi via ssh
-    ssh <user>@<picarx-ip>
+# Login to the RaspberryPi via ssh
+ssh <user>@<picarx-ip>
 
-    # Build and execute the Docker Containers
-    TAG=latest ARCH=arm32v7 docker compose -f docker-compose-core.yml build 
-    TAG=latest ARCH=arm32v7 docker compose -f docker-compose-pt.yml build 
-    TAG=latest ARCH=arm32v7 docker compose -f docker-compose-pt.yml up  {% endhighlight %}
+# Build and execute the Docker Containers
+TAG=latest ARCH=arm32v7 docker compose -f docker-compose-core.yml build 
+TAG=latest ARCH=arm32v7 docker compose -f docker-compose-pt.yml build 
+TAG=latest ARCH=arm32v7 docker compose -f docker-compose-pt.yml up  {% endhighlight %}
   </div>
   <div class="tab-content" id="tab3-2">
   {% highlight console %}
-    # First copy all content to the RaspberryPi4
-    scp -r ./ <user>@<picarx-ip>:~/
+# First copy all content to the RaspberryPi4
+scp -r ./ <user>@<picarx-ip>:~/
 
-    # Login to the RaspberryPi via ssh
-    ssh <user>@<picarx-ip>
+# Login to the RaspberryPi via ssh
+ssh <user>@<picarx-ip>
 
-    # Build and execute the Docker Containers
-    TAG=latest ARCH=arm64v8 docker compose -f docker-compose-core.yml build 
-    TAG=latest ARCH=arm64v8 docker compose -f docker-compose-pt.yml build 
-    TAG=latest ARCH=arm64v8 docker compose -f docker-compose-pt.yml up  {% endhighlight %}  
+# Build and execute the Docker Containers
+TAG=latest ARCH=arm64v8 docker compose -f docker-compose-core.yml build 
+TAG=latest ARCH=arm64v8 docker compose -f docker-compose-pt.yml build 
+TAG=latest ARCH=arm64v8 docker compose -f docker-compose-pt.yml up  {% endhighlight %}  
   </div>
 </div>
 
[…]
