+++ docs/getting_started.md
@@ -10,15 +10,15 @@ nav_order: 2
 # Getting Started
 This project based on ROS and Docker. Due to the used interfaces on the RPi, we have to use Linux Kernel functions for GPIO and I2C. Before you can start this project you have to activate GPIO and I2C. If you already activated these modules, you can proceed with ##. Otherwise you have to build these modules first.
 
-| The PiCar-X by Sunfounder                                            | The digital twin prototype in GAZEBO                                                               |
+| The PiCar-X by Sunfounder| The digital twin prototype in GAZEBO   |
 | -------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------- |
 | ![Physical Twin](./assets/images/picarx-pt.jpg "The PiCar-x by Sunfounder") | ![Digital Twin Prototype](./assets/images/picarx-gazebo.gif "The digital twin prototype of the PiCar-X") |
 
 
 ## Clone Repository
 ```console 
-    git clone https://github.com/cau-se/ARCHES-PiCar-X.git
-    cd ./ARCHES-PiCar-X/PiCar-X
+git clone https://github.com/cau-se/ARCHES-PiCar-X.git
+cd ./ARCHES-PiCar-X/PiCar-X
 ```
 
 ## Check if required modules are installed
@@ -50,18 +50,18 @@ If you already have built the modules, you need to activate them via:
 
 <div class="tab-container" id="activateinterfaces">
   <ul class="tab-list">
-    <li class="tab active" data-tab="tab1-1">X64</li>
-    <li class="tab" data-tab="tab1-2">RaspberryPi 3/4</li>
+<li class="tab active" data-tab="tab1-1">X64</li>
+<li class="tab" data-tab="tab1-2">RaspberryPi 3/4</li>
   </ul>
   <div class="tab-content active" id="tab1-1">
   {% highlight console %}
-    sudo modprobe gpio-mockup gpio_mockup_ranges=1,41
-    sudo modprobe i2c-dev
-    sudo modprobe i2c-stub chip_addr=0x14 {% endhighlight %}
+sudo modprobe gpio-mockup gpio_mockup_ranges=1,41
+sudo modprobe i2c-dev
+sudo modprobe i2c-stub chip_addr=0x14 {% endhighlight %}
   </div>
   <div class="tab-content" id="tab1-2">
   {% highlight console %}
-    sudo modprobe i2c-stub chip_addr=0x14 {% endhighlight %}  
+sudo modprobe i2c-stub chip_addr=0x14 {% endhighlight %}  
   </div>
 
 </div>
@@ -73,6 +73,27 @@ ls /dev/i2c*
 
 If there was already and existing I2C device (see above), then a new would should be listed now. The result can be something like `/dev/i2c-0 /dev/i2c-1` and `/dev/i2c-1` is the created device that can be used for the DTP.
 
+## Enable Gazebo GUI
+Before starting the Docker containers, enable the X11 forwarding to the hosts display:
+
+```console
+# Alternative 1: Allow all connetions
+xhost + 
+
+# Alternative 2: Allow non-network local connections
+xhost + local:root
+```
+
+**DEACTIVATE AFTER SHUTDOWN**
+```console
+# Alternative 1: Disable all connetions
+xhost - 
+
+# Alternative 2: Disable non-network local connections
[…]

[…]
