### 2024-09-10 a69ca2d
adjusted docker compose to gazebo gui with x11 on ubuntu #4
