+++ PiCar-X/unexport-gpio.sh
@@ -1,0 +1,2 @@
+echo 23 > /sys/class/gpio/unexport
+echo 24 > /sys/class/gpio/unexport
