### 2024-09-10 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
Sorry for the late response. If I understand the comments correct, the I2C problem is solved. Thanks for posting the results, I will place it on the troubleshooting page.

In the end, the installation guide for the modules was correct, but the configuration with the I2C variable in the environemnt files was missing, right?

**Visualization with a Gazebo Simulation**
I adjusted the docker compose file to use xhost for the Gazebo GUI on both, Ubuntu and WSL2. Therefore, you have to install:
```console
sudo apt install xorg
```

Afterwards, you can activate the X11 forwarding via
```console
xhost +
```

And then you can start the docker compose file. I tested this with Ubuntu and WSL2 and it worked on both systems.

After stopping the docker compose file remove the X11 forwarding again.

```console
xhost -
```

This is no perfect solution, but I was not able to limit it to the user level. I added this part also to the documentation.

**Visualization in a terminal**
An alternative way to display the content without the GUI would be to switch into a container

```console
docker exec -it picar-x-ackermann_skill-dtp-1 /bin/bash

# in general, this should not be required, but in case the .bashrc is not loaded..
source ./install/setup.bash

# get all available topics with
rostopic list

# enter any of the listed topics after rostopic echo
rostopic echo /drive/status
```

In a second terminal you also switch into a container, but there you publish a message:

```console
docker exec -it picar-x-ackermann_skill-dtp-1 /bin/bash

# in general, this should not be required, but in case the .bashrc is not loaded..
source ./install/setup.bash

rostopic pub /picarx/drive/command picarx_msgs/Drive "{speed: 50, angle: 20}"
```

With regard to your other remarks:
- I removed the whitespaces
- The command was not a typo, but the active tab and the tab content were mixed up, fixed it.
