### 2024-09-10 e74e497
adjusted i2c detection in docu #4
