### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
At the moment I have to define the environment variable I2C with

```I2C=/dev/i2c-3  docker compose -f docker-compose-dtp.yml up```

It does not work without it.

I also have to activate the modules after each reboot of the host system:

```
sudo modprobe gpio-mockup gpio_mockup_ranges=1,41
sudo modprobe i2c-dev
sudo modprobe i2c-stub chip_addr=0x14
```

I will test the GUI in a minute.
