### 2024-09-11 cau-se/ARCHES-PiCar-X#6 @AlexanderFabisch
Documentation: Tabs don't initialize correctly under "Testing"
When I go to the page ["Testing"](https://cau-se.github.io/ARCHES-PiCar-X/integration_tests.html) and scroll down to "Manual Integration Testing", the first tab "The full DTP on a X64 machine" (should probably be "an X64..." since X is pronounced "ex"?) shows

```
# Build the core container
TAG=latest-arm32v7 ARCH=arm32v7 docker compose -f docker-compose-core.yml build --no-cache

...
```

When I switch to another tab and then switch back to the same tab again then it shows correctly

```
# Build the core container
TAG=latest docker compose -f docker-compose-core.yml build --no-cache

...
```

openjournals/joss-reviews/issues/7179
