### 2024-09-11 cau-se/ARCHES-PiCar-X#7 @AlexanderFabisch
Documentation: error in image name for integration tests
On my machine the name of the container to run the integration tests is

`ghcr.io/cau-se/arches-picar-x/drivers/dcmotor`

In the documentation it says

`ghcr.io/cau-searches-picar-x/drivers/dcmotor`

So the slash after cau-se is missing.

openjournals/joss-reviews/issues/7179
