### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
> And then you can start the docker compose file. I tested this with Ubuntu and WSL2 and it worked on both systems.

Do I have to change anything else to start the Gazebo GUI? Because it doesn't work like that on my machine after running `xhost +` and also rebuilding the images.

Everything else works as intended now. Thanks a lot! I apologize for being a bit helpless here since I never worked with this before. But I think the documentation improved for other clueless researchers. :)
