### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
>Do I have to change anything else to start the Gazebo GUI? Because it doesn't work like that on my machine after running xhost + and also rebuilding the images.

Have you pulled the latest commits? I changed the environmental variable display and mounted /tmp/X.11-unix into the container. On my systems that was enough.

Instead of starting via docker compose, you could try
```console
docker run --rm --name guitest -i -e DISPLAY -v /tmp/X.11-unix:/tmp/X.11-unix --net=host ghcr.io/cau-se/arches-picar-x/picarx-gazebo::latest roslaunch picarx_gazebo picarx-world.launch
```
If this opens, there is still some problem in the compose file.

> Everything else works as intended now. Thanks a lot! I apologize for being a bit helpless here since I never worked with this before. But I think the documentation improved for other clueless researchers. :)

Thank you very much for you patience! :) refactored the compose files a few weeks ago to allow the environmental variables instead of changing each compose file individually. But quite error prone this approach.
