### 2024-09-11 5f3ea79
added interactive shell picarx-gazebo container in dtp compose file #4
