### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
Okay, maybe the interactive shell activation in the compose file was missing. Did not affect my WSL setup, but for ubuntu this seems important.
