### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
This one works:
> ```
> docker run --rm --name guitest -i -e DISPLAY -v /tmp/X.11-unix:/tmp/X.11-unix --net=host ghcr.io/cau-se/arches-picar-x/picarx-gazebo::latest roslaunch picarx_gazebo picarx-world.launch
> ```

I pulled from Github, rebuild the docker images and used docker compose to start everything, but the Gazebo GUI doesn't show up. But there is an error message. Unfortunately I was not able to copy the log file mentioned to the host.

```
picarx-gazebo-1             | started roslaunch server http://picarx-gazebo:37623/
picarx-gazebo-1             |
picarx-gazebo-1             | SUMMARY
picarx-gazebo-1             | ========
picarx-gazebo-1             |
picarx-gazebo-1             | PARAMETERS
picarx-gazebo-1             |  * /gazebo/enable_ros_network: True
picarx-gazebo-1             |  * /robot_description: <?xml version="1....
picarx-gazebo-1             |  * /rosdistro: noetic
picarx-gazebo-1             |  * /rosversion: 1.16.0
picarx-gazebo-1             |  * /use_sim_time: True
picarx-gazebo-1             |
picarx-gazebo-1             | NODES
picarx-gazebo-1             |   /
picarx-gazebo-1             |     gazebo (gazebo_ros/gzserver)
picarx-gazebo-1             |     gazebo_gui (gazebo_ros/gzclient)
picarx-gazebo-1             |     spawn_urdf (gazebo_ros/spawn_model)
picarx-gazebo-1             |
picarx-gazebo-1             | auto-starting new master
picarx-gazebo-1             | process[master]: started with pid [34]
picarx-1                    | ROS_MASTER_URI=http://picarx:11311/
picarx-1                    |
picarx-1                    | setting /run_id to 34624818-7011-11ef-b02b-0242ac120002
picarx-1                    | process[rosout-1]: started with pid [42]
picarx-1                    | started core service [/rosout]
picarx-gazebo-1             | ROS_MASTER_URI=http://picarx-gazebo:11312
picarx-gazebo-1             |
picarx-gazebo-1             | setting /run_id to 346503be-7011-11ef-b0a5-0242ac120003
picarx-gazebo-1             | process[rosout-1]: started with pid [44]
picarx-gazebo-1             | started core service [/rosout]
picarx-gazebo-1             | process[gazebo-2]: started with pid [47]
picarx-gazebo-1             | process[gazebo_gui-3]: started with pid [54]
picarx-gazebo-1             | process[spawn_urdf-4]: started with pid [61]
picarx-gazebo-1             | [ INFO] [1726040467.581318713]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1726040467.582130798]: waitForService: Service [/gazebo/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1726040467.652530124]: Finished loading Gazebo ROS API Plugin.
picarx-gazebo-1             | [ INFO] [1726040467.653397659]: waitForService: Service [/gazebo_gui/set_physics_properties] has not been advertised, waiting...
picarx-gazebo-1             | [ INFO] [1726040467.725197957]: waitForService: Service [/gazebo/set_physics_properties] is now available.
picarx-gazebo-1             | [ INFO] [1726040467.736731687, 0.010000000]: Physics dynamic reconfigure ready.
picarx-gazebo-1             | Aborted (core dumped)
picarx-gazebo-1             | [gazebo_gui-3] process has died [pid 54, exit code 134, cmd /opt/ros/noetic/lib/gazebo_ros/gzclient __name:=gazebo_gui __log:=/root/.ros/log/346503be-7011-11ef-b0a5-0242ac120003/gazebo_gui-3.log].
picarx-gazebo-1             | log file: /root/.ros/log/346503be-7011-11ef-b0a5-0242ac120003/gazebo_gui-3*.log
picarx-gazebo-1             | [ INFO] [1726040468.026205265, 0.177000000]: Loading gazebo_ros_control plugin
picarx-gazebo-1             | [ INFO] [1726040468.026284914, 0.177000000]: Starting gazebo_ros_control plugin in namespace: /picarx
picarx-gazebo-1             | [ INFO] [1726040468.026668431, 0.177000000]: gazebo_ros_control plugin is waiting for model URDF in parameter [/robot_description] on the ROS param server.
picarx-gazebo-1             | [ERROR] [1726040468.169133224, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_right
picarx-gazebo-1             | [ERROR] [1726040468.170563922, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/joint_wheel_left
picarx-gazebo-1             | [ERROR] [1726040468.171776769, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_left
picarx-gazebo-1             | [ERROR] [1726040468.172886364, 0.177000000]: No p gain specified for pid.  Namespace: /picarx/gazebo_ros_control/pid_gains/steer_right
picarx-gazebo-1             | [ INFO] [1726040468.176124421, 0.177000000]: Loaded gazebo_ros_control.
picarx-gazebo-1             | [spawn_urdf-4] process has finished cleanly
picarx-gazebo-1             | log file: /root/.ros/log/346503be-7011-11ef-b0a5-0242ac120003/spawn_urdf-4*.log
```
