### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
The GUI process gets killed in the compose process. But if it is working without docker compose, we are not far away from our goal.

In my last commit I added
`stdin_open: true` to the picarx-gazebo container in the `docker-compose-dtp.yml`, did you also pull this commit?
