### 2024-09-11 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
> In my last commit I added
> stdin_open: true to the picarx-gazebo container in the docker-compose-dtp.yml, did you also pull this commit?

Yes, and then rebuilt the images.
