### 2024-09-12 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
> hmmmm, so I cloned the repo again on my Ubuntu system, activate xhost with `xhost +`, adjusted the I2C in the `simulation.env` and `picarx.env`, built the `docker-compose-core.yml` and the `docker-compose-dtp.yml`, started it with docker compose and the GUI opened. I will set up a new Ubuntu System tomorrow evening and try it again step by step, but have to leave now for a business trip.

I tried the same. Didn't work on my machine.

> I do not know if xhost could block something with the docker user, maybe try `xhost + local:root` or `xhost local:docker`, but if the GUI opens with plain docker run..

Tried that too.

> Do you use the plain docker engine on your Ubuntu or did you install the Docker Desktop application?

I did not install the desktop application.

> **EDIT:** what I was just thinking about, try to rebuild the containers with `--no-cache`:

I've done that multiple times, but nothing seems to work.
