### 2024-09-12 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
I reinstalled the Ubuntu two times now and copied the instructions fromt he gruide to build the kernel. What i am doing different from the guide is that I used the gui tool `grub-customizer` to switch from kernel 5.15.0-119-generic (default in the Ubuntu 20.04 Image on the official ubuntu page) to kernel 5.13.0. Furthermore, I installed the packages `git` and `curl`. For Docker I used the official Docker installation guide. I will add git and curl in the dependency list, reference to the docker guide and add the grub-customizer to the troubleshooting page, in case someone wants to switch easily between kernel versions.

However, the Gazebo window always opens after `xhost +` and the docker compose. I am wondering whether the grub-customizer installs dependencies that can influence the behavior. Can you please try:

```console
sudo apt install grub-customizer
```

And then start the docker compose again.
