### 2024-09-12 b9f83d5
a $ was missing in the roslaunch command of the steering-emulator that prevented the bash variable replacment
