### 2024-09-12 1bf1476
added git and curl to dependencies, added -y flag to auto accept installation, and mentioned docker installation guide #4
