### 2024-09-13 cau-se/ARCHES-PiCar-X#4 @AlexanderFabisch
I almost gave up. Then I checked the docker compose file again. It works now on my machine! This is the git diff. I can see the picar moving! :)

```
diff --git a/PiCar-X/docker-compose-dtp.yml b/PiCar-X/docker-compose-dtp.yml
index 67c6ef6..be6db24 100644
--- a/PiCar-X/docker-compose-dtp.yml
+++ b/PiCar-X/docker-compose-dtp.yml
@@ -151,7 +151,7 @@ services:
     ports:
       - 11312:11312
     volumes:
-      - ${X11PATH:-/tmp/.X11-unix/X0}:/tmp/.X11-unix/X0
+      - ${X11PATH:-/tmp/.X11-unix}:/tmp/.X11-unix
       - ./core:/root/catkin_ws/src/core
       - ./simulation:/root/catkin_ws/src/simulation
     command: /bin/bash -c "roslaunch picarx_gazebo picarx-world.launch"
```
