### 2024-09-15 2e833ff
adjusted compose files to running config #4
