+++ PiCar-X/docker-compose-dtsim.yml
@@ -25,31 +25,6 @@ services:
       retries: 5
       start_period: 1s
 
-  picarx-gazebo:
-    image: ghcr.io/cau-se/arches-picar-x/picarx-gazebo:${TAG:-latest}
-    build:
-      context: Dockerfile
-      dockerfile: ./simulation
-      target: desktop
-      args:
-        - TAG=${TAG:-latest}
-    env_file:
-      - ./env/simulation.env
-    environment:
-      - DISPLAY=host.docker.internal:0.0
-      - LIBGL_ALWAYS_INDIRECT=0
-      - ROS_HOSTNAME=picarx-gazebo
-    tty: true
-    ports:
-      - 11312:11312
-    volumes:
-      - ./core:/root/catkin_ws/src/core
-      - ./simulation:/root/catkin_ws/src/simulation
-    command: /bin/bash -c "roslaunch picarx_gazebo picarx-world.launch & roslaunch picarx_control picarx_control.launch"
-    depends_on:
-      picarx-dt:
-        condition: service_healthy
-
   shadow-dt:
     image: ghcr.io/cau-se/arches-picar-x/picarx:${TAG:-latest}
     env_file:
@@ -71,7 +46,7 @@ services:
     env_file:
       - ./env/picarx.env
     tty: true
-    command: /bin/bash -c "roslaunch arches_mqtt_bridge dt_client.launch host:=mqtt-server"
+    command: /bin/bash -c "roslaunch arches_mqtt_bridge dt_client.launch host:=mqtt-server  motor1_i2c_port:=$${I2C:-/dev/i2c-0}"
     depends_on:
       picarx-dt:
         condition: service_healthy
@@ -93,7 +68,7 @@ services:
       - ${I2C:-/dev/i2c-0}:${I2C:-/dev/i2c-0}
       - ./core:/root/catkin_ws/src/core
       - ./ros/drivers/dcmotor:/root/catkin_ws/src/drivers/dcmotor
-    command: /bin/bash -c "roslaunch picarx_dcmotor_driver dcmotor_left.launch"
+    command: /bin/bash -c "roslaunch picarx_dcmotor_driver dcmotor_left.launch  motor2_i2c_port:=$${I2C:-/dev/i2c-0}"
     depends_on:
       picarx-dt:
         condition: service_healthy
@@ -131,7 +106,7 @@ services:
       - ${I2C:-/dev/i2c-0}:${I2C:-/dev/i2c-0}
       - ./core:/root/catkin_ws/src/core
       - ./ros/drivers/clutchgear:/root/catkin_ws/src/drivers/clutchgear
-    command: /bin/bash -c "roslaunch picarx_clutchgear_driver ackermann_clutchgear_driver.launch"
+    command: /bin/bash -c "roslaunch picarx_clutchgear_driver ackermann_clutchgear_driver.launch i2c_port:=$${I2C:-/dev/i2c-0}"
     depends_on:
       picarx-dt:
         condition: service_healthy
@@ -158,6 +133,50 @@ services:
       picarx-dt:
         condition: service_healthy
 
+  picarx-gazebo:
+    image: ghcr.io/cau-se/arches-picar-x/picarx-gazebo:${TAG:-latest}
+    build:
+      context: Dockerfile
+      dockerfile: ./simulation
+      target: desktop
+      args:
+        - TAG=${TAG:-latest}
+    env_file:
+      - ./env/simulation.env
+    environment:
+      - DISPLAY=$DISPLAY
+      - LIBGL_ALWAYS_INDIRECT=0
+      - ROS_HOSTNAME=picarx-gazebo
+    stdin_open: true
+    tty: true
+    ports:
[…]
