### 2024-09-16 ffd3462
added statement of need to landing page and summarizing the different dt definitions
