+++ docs/index.md
@@ -5,10 +5,23 @@ nav_order: 1
 permalink: /
 ---
 
+While digital twins are becoming increasingly important in fields like Industry 4.0 and the Internet of Things (IIoT), there is a significant gap in terms of reproducibility and validation in current digital twin research.
+Despite the increasing interest in digital twin technology, many existing studies fail to provide detailed, reusable, and reproducible specifications and source code, making it difficult for researchers and engineers to replicate or extend these concepts independently.
+
+We see a clear need for affordable, accessible, and reproducible laboratory examples that other researchers can use to explore and validate digital twin implementations. To meet this need, we present the ARCHES PiCar-X, where we implement various digital twin concepts. By providing open-source code and Docker-compose files, we aim to enable independent replication and extension, thus advancing the reproducibility of digital twin research. Reproducibility is essential for scientific progress, and we seek to fill the gap in reproducible digital twin examples available to the research community.
+
 The ARCHES PiCar-X project can be used to explore the digital twin concept developed and employed in the project ARCHES (Autonomous Robotic System to help Human Societies). 
 {: .fs-6 .fw-300 .text-center } 
 
 [Get Started Now](https://cau-se.github.io/ARCHES-PiCar-X/getting_started.html){: .btn .btn-primary .fs-5 .mb-4 .mb-md-0 .mr-4 }
 {: .text-center }
+
+This example aims to demonstrate the different digital concepts we formally specified in:
+
+>Barbie, A., & Hasselbring, W. (2024). From Digital Twins to Digital Twin Prototypes: Concepts, Formalization, and Applications. IEEE Access. [https://doi.org/10.1109/access.2024.3406510](https://doi.org/10.1109/access.2024.3406510)
+
+How the different concepts are implemented with the ARCHES PiCar-X is described in:
+
+>Barbie, A., & Hasselbring, W. (2024). Toward Reproducibility of Digital Twin Research: Exemplified with the PiCar-X. arXiv preprint arXiv:2408.13866. [https://doi.org/10.48550/ARXIV.2408.13866](https://doi.org/10.48550/ARXIV.2408.13866)
 
 ---
