### 2024-09-16 94e83a1
Removed duplicate contributing headline
