+++ README.md
@@ -12,12 +12,12 @@ This repository presents the source code for a digital twin prototype of the [Pi
 
 The ARCHES PiCar-X can be run via Docker. Follow our [quickstart guide](https://cau-se.github.io/ARCHES-PiCar-X/getting_started.html) to get started.
 
-## Documentation
+# Documentation
 
 Documentation on The ARCHES PiCar-X and additional information can be found on our [`GitHub Page`](https://cau-se.github.io/ARCHES-PiCar-X/)
 
 
-## Contributing
+# Contributing
 
 We are happy to accept any kind of contributions to the ARCHES PiCar-X.
 
@@ -30,17 +30,15 @@ See our website to [start contributing](https://cau-se.github.io/ARCHES-PiCar-X/
 * The corresponding paper is placed in [`paper`](paper)
 * The documentation can be found on our [`GitHub Page`](https://cau-se.github.io/ARCHES-PiCar-X/)
 
-# Contributing
-
 # Acknowledgements
 We thank the GitHub user [Theosakamg](https://github.com/Theosakamg) for providing a CAD model of a PiCar-V under an open source license.
 
 # Related Publications
-- Barbie, A., Hasselbring, W., Pech, N., Sommer, S., Flögel, S., & Wenzhöfer, F. (2020, September). Prototyping autonomous robotic networks on different layers of rami 4.0 with digital twins. In 2020 IEEE International Conference on Multisensor Fusion and Integration for Intelligent Systems (MFI) (pp. 1-6). IEEE. [https://doi.org/10.1109/mfi49285.2020.9235210](https://doi.org/10.1109/mfi49285.2020.9235210)
-- Barbie, A., Pech, N., Hasselbring, W., Flögel, S., Wenzhöfer, F., Walter, M., ... & Sommer, S. (2021). Developing an underwater network of ocean observation systems with digital twin prototypes—a field report from the baltic sea. IEEE Internet Computing, 26(3), 33-42.  [https://doi.org/10.1109/mic.2021.3065245](https://doi.org/10.1109/mic.2021.3065245)
+- Barbie, A., Hasselbring, W., Pech, N., Sommer, S., Flögel, S., & Wenzhöfer, F. (2020, September). Prototyping Autonomous Robotic Networks on Different Layers of RAMI 4.0 with Digital Twins. In 2020 IEEE International Conference on Multisensor Fusion and Integration for Intelligent Systems (MFI) (pp. 1-6). IEEE. [https://doi.org/10.1109/mfi49285.2020.9235210](https://doi.org/10.1109/mfi49285.2020.9235210)
+- Barbie, A., Pech, N., Hasselbring, W., Flögel, S., Wenzhöfer, F., Walter, M., ... & Sommer, S. (2021). Developing an Underwater Network of Ocean Observation Systems with Digital Twin Prototypes - A Field Report from the Baltic Sea. IEEE Internet Computing, 26(3), 33-42.  [https://doi.org/10.1109/mic.2021.3065245](https://doi.org/10.1109/mic.2021.3065245)
 - Barbie, A., & Pech, N. (2022). ARCHES Digital Twin Framework. GEOMAR Helmholtz Centre for Ocean Research Kiel. [https://doi.org/10.3289/sw_arches_core_1.0.0](https://doi.org/10.3289/sw_arches_core_1.0.0)
[…]
