### 2024-09-16 36c9165
adjusted definitions to formalization paper
