+++ docs/digitaltwins/overview.md
@@ -20,84 +20,86 @@ The ARCHES PiCar-X was specifically developed to help scientists and engineers u
 ## 1. Physical Twin
 
 ### Definition
-A **Physical Twin** is the real-world physical system or product that a digital twin is based on. It consists of the physical object itself, which is equipped with sensors or actuators and controlled by embedded software.
+A **physical twin** is a real-world physical System-of-Systems or product. It comprises sensing or actuation capabilities driven by embedded software.
 
 ### Key Points:
 - A real object, such as a vehicle or machine.
 - Equipped with sensors and actuators for interaction with its environment.
-- In the context of the PiCar-X, the **Physical Twin** is the PiCar-X itself, with all its hardware components, including motors, sensors, and embedded software.
+- In the context of the PiCar-X, the **physical twin** is the PiCar-X itself, with all its hardware components, including motors, sensors, and embedded software.
 
 ---
 
 ## 2. Digital Model
 
 ### Definition
-A **Digital Model** is a virtual representation of the physical object, process, or system. It can be a mathematical model or a computer-aided design (CAD) used to simulate the behavior of the physical twin.
+A **digital model** describes an object, a process, or a complex aggregation. The description is either a mathematical or a computeraided design (CAD).
 
 ### Key Points:
 - Provides a virtual, often simplified, version of the physical twin.
 - Can be used for simulations and testing without interacting with the actual hardware.
-- For the PiCar-X, a **CAD model** is used in a GAZEBO simulation, representing the car's dimensions and driving behavior.
+- For the PiCar-X, a CAD model is used in a GAZEBO simulation, representing the car's dimensions and driving behavior.
 
 ---
 
 ## 3. Digital Template
 
 ### Definition
-A **Digital Template** serves as a framework that can be populated with specific data to generate a digital or physical twin. It includes the necessary software, models, and documentation for replicating or understanding the system.
+A **digital template** serves as a framework that can be tailored or populated with specific information to generate the physical twin. It encompasses the software operating the physical twin, its digital model,
+and all the essential information needed for constructing and sustaining the physical twin, such as blueprints, bills of materials, technical manuals, and similar documentation.
 
 ### Key Points:
 - Combines the physical twin’s software, digital models, blueprints, and documentation.
 - Provides the structure and data needed to recreate or understand the physical twin.
-- In the case of the PiCar-X, the **Digital Template** includes software, a CAD model, and documentation on materials and schematics.
+- In the case of the PiCar-X, the **digital template** includes software, a CAD model, and documentation on materials and schematics.
 
 ---
 
 ## 4. Digital Thread
 
 ### Definition
-A **Digital Thread** refers to the communication framework that enables the flow of data between the physical twin and its digital representations throughout the product's lifecycle.
+The **digital thread** refers to the communication framework that allows a connected data flow and integrated view of the physical twin’s data and operations throughout its lifecycle.
 
 ### Key Points:
-- Ensures continuous data flow between physical and digital components.
-- Provides real-time data on the physical twin's status to the digital twin.
-- In the PiCar-X example, the **Digital Thread** uses ROS (Robot Operating System) and the MQTT protocol to connect the physical PiCar-X to its digital model and synchronize data.
+- Ensures continuous data flow between physical and digital components, but is not limited to this communication.
+- Enables (near) real-time monitoring of the physical twin's status on the digital twin.
+- In the PiCar-X example, the **digital thread** uses ROS (Robot Operating System) and the MQTT protocol to connect the physical PiCar-X to its digital model and synchronize data.
 
 ---
 
 ## 5. Digital Shadow
 
 ### Definition
[…]
