### 2024-09-19 cau-se/ARCHES-PiCar-X#8 @mrsonandrade
Error during building
**Describe the bug**
Everything was fine following the "Getting Started". Thus, as I don't have the physical car, I followed to “Explore -> Example for replacing the physical twin with a digital twin prototype  -> Option 3: Only the Digital Twin starts a simulation”. Thus, for this line:

`docker compose -f docker-compose-dtsim.yml build`

I got this error during the building:
`service “motor-emulator_right-dt” depends on undefined service “picarx-dt”: invalid compose project`

**To Reproduce**
Steps to reproduce the behavior:
1. After success with the [Getting Started](https://cau-se.github.io/ARCHES-PiCar-X/getting_started.html), go to [Replace the Physical Twin with a Digital Twin Prototype and Connect it to a Digital Twin](https://cau-se.github.io/ARCHES-PiCar-X/digitaltwins/dtp.html).
2. Follow the [Option 3](https://cau-se.github.io/ARCHES-PiCar-X/digitaltwins/dtp.html#option3).
3. When you give `docker compose -f docker-compose-dtsim.yml build`.
4. You have the following error `service “motor-emulator_right-dt” depends on undefined service “picarx-dt”: invalid compose project`.

**Additional context**
Before all steps above I tested the [Let the ARCHES PiCar-X Drive](https://cau-se.github.io/ARCHES-PiCar-X/getting_started.html#let-the-arches-picar-x-drive) and everything was working great.

https://github.com/openjournals/joss-reviews/issues/7179
