### 2024-09-19 730ddb7
some additions to the paper
