+++ paper/paper.md
@@ -23,35 +23,32 @@ bibliography: paper.bib
 ---
 
 # Summary
-This paper presents a digital twin prototype of the [PiCar-X by Sunfounder](https://www.sunfounder.com/products/picar-x) based on the Robot Operating System (ROS) and the ARCHES Digital Twin Framework. Our goal is to provide researchers and practitioners with an affordable and straightforward example to explore various concepts of digital twins. Moreover, this example can be used to explore all the concepts implemented in a previous project named ARCHES (Autonomous Robotic Networks to help Human Societies), which focused on monitoring and operating an underwater network of ocean observation systems. This network was successfully tested in the Baltic Sea in October 2020. One of the outcomes of this project was the ARCHES Digital Twin Framework.
+This paper presents a digital twin prototype of the [PiCar-X by Sunfounder](https://www.sunfounder.com/products/picar-x) based on the middleware [Robot Operating System (ROS)](https://ros.org), Docker, and the ARCHES Digital Twin Framework, which provides tools to exchange data between a physical and digital twin. The digital twin prototype can be used to test all software functions without needing to use the physical PiCar-X. The actual hardware is replaced with emulators or a simulation, and the interfaces are virtualized. Our goal is to provide researchers and practitioners with an affordable and straightforward example to explore how the concepts physical twin, digital model, digital template, digital thread, digital shadow, digital twin, and digital twin prototype can be implemented. These concepts were originally used for development, testing, monitoring, and operating an underwater network of ocean observation systems in the project ARCHES (Autonomous Robotic Networks to help Human Societies).
 
 # Statement of need
-Digital twins are becoming increasingly relevant in the Industrial Internet of Things and Industry 4.0, enhancing the capabilities and quality of various applications. However, the concept of digital twins lacks a unified definition and faces validation challenges, partly due to the scarcity of reproducible modules or source codes in existing studies. While many applications are described in case studies, they often lack detailed, re-usable specifications for researchers and engineers.
-In previous research, we developed a network of ocean observation systems and evaluated the results in a real-world mission in the Baltic sea in Oktober 2020 [@demomission:2020]. One of the results of the successful proof-of-concept was the ARCHES Digital Twin Framework [@ADTF:2022], a software package providing the functionality to implement the digital thread between physical twins and digital twins.
-Ocean observation systems use quite specific and expensive hardware, hence, we see the need of a cheap lab experiemnt to enable independent evaluation and exploration of the different concepts. The PiCar-X example demonstrates all the concepts from the ocean observation system and includes also a full integration test pipeline, see our GitHub Repository [@archespicarx:2024].
+Digital twins are becoming increasingly relevant in the Industrial Internet of Things and Industry 4.0, enhancing the capabilities and quality of various applications [@kritzinger:2018]. However, the concept of digital twins lacks a unified definition and faces validation challenges, partly due to the scarcity of reproducible modules or source codes in existing studies. While many applications are described in case studies, they often lack detailed, re-usable specifications for researchers and engineers.
 
-The PiCar-X can also be used to evaluate the different concepts, namely physical twin, digital model, digital template, digital thread, digital shadow, digital twin, and digital twin prototype, which we formally specified in [@barbie:2024] using the Object-Z notation. In [@barbiepicarx:2024], we elaborate in more detail how the PiCar-X can be used to evaluate all these concepts.
[…]
