+++ paper/paper.bib
@@ -88,3 +88,15 @@ url = {https://github.com/acmsigsoft/EmpiricalStandards}
   journal = {GitHub repository},
   url = {https://github.com/cau-se/ARCHES-PiCar-X}
 }
+
+@article{kritzinger:2018,
+  doi = {10.1016/j.ifacol.2018.08.474},
+  year = {2018},
+  publisher = {Elsevier {BV}},
+  volume = {51},
+  number = {11},
+  pages = {1016--1022},
+  author = {Werner Kritzinger and Matthias Karner and Georg Traar and Jan Henjes and Wilfried Sihn},
+  title = {{Digital Twin in manufacturing: A categorical literature review and classification}},
+  journal = {{IFAC}-{PapersOnLine}}
+}
