### 2024-09-19 2792bf6
renamed roscore container and using picarx-dt.env #8
