### 2024-09-19 69caec1
fixed and tested refactoring errors in compose files #8
