### 2024-09-19 6d927cb
adjusted docs for DPT-DT setup
