### 2024-09-19 cau-se/ARCHES-PiCar-X#8 @AlexanderBarbie
Great news that the installation worked. :)

I fixed the errors in the compose files and they build and start on my systems now. Should have tested this better after refactoring...

Since you were already on the explore page: I adjusted the documentation for the setup with the DTP including Gazebo and the DT including Gazebo (option 1). I started it on the same system, but was never thinking about the fact that they both share the same I2C device and i cannot create a second I2C-stub device on my system. I will elaborate this part in the documentation at the weekend.
