### 2024-09-25 cau-se/ARCHES-PiCar-X#4 @danielskatz
Can this be closed now? (sorry if not, this is a long thread...)
