### 2024-09-25 #7179 @danielskatz
Hi All - it looks like things are moving along on this, which is good.

@AlexanderBarbie - are you working on the comments above (https://github.com/openjournals/joss-reviews/issues/7179#issuecomment-2343629103)?
