### 2024-09-26 cau-se/ARCHES-PiCar-X#4 @AlexanderBarbie
I was not sure if we maybe still need this issue, but it does not look like. I will create a new issue for the documentation and enhance the documentation at the weekend and also rewrite the paper a little bit.
