### 2024-09-27 7a5924c
adjustments to paper
