+++ paper/paper.md
@@ -32,8 +32,10 @@ In [@barbie:2024], we formally specified a digital twin concept including its su
 Ocean observation systems use quite specific and expensive hardware, hence, we see the need of a cheap lab experiemnt to enable independent evaluation and exploration of the different concepts. The PiCar-X example demonstrates all the concepts from the ocean observation system and includes also a full integration test pipeline, see our GitHub Repository [@archespicarx:2024]. In [@barbiepicarx:2024], we elaborate in more detail how the PiCar-X can be used to evaluate all these concepts.
 
 
-# The PiCar-X and its Digital Model
-The PiCar-X is a toy car, see \autoref{fig:picarx-pt}, with all sensors and actuators connected to a RaspberryPi. Two direct current motors (DC motors) are used to move the car. A servo motor at the front is used to steer the car. The steering is a typical Ackermann steering [@Veneri:2020] known from common cars. The PiCar-X also includes grey-scale sensors for line following, infrared sensors for collision avoidance, and a camera. However, in the current example only the DC motors and the servo motor for steering are included, so far. All software components are implemented using the ROS middleware and containerized in Docker. The ARCHES Digital Twin Framework establishes the digital thread between the physical twin and the digital shadow/twin [@barbiepicarx:2024].
+# The PiCar-X and its Digital Twin
+There is no clear definition of a digital twin. The range of interpretations spans from a complex simulation to a completely mirrored status of the physical device, which can also be controlled via the digital twin. In our definition [@barbie:2024], the digital twin is connected to the physical twin over the entire life cycle for automated bidirectional data exchange, i.e. changes made to the digital twin lead to adapted behavior of the physical twin and vice-versa. The implementation of this capabilties can very, with model driven approachen are being very popular [@barbie:2024]. Hence, we present a lab experiment that can be used to explore our approach independely. 
+
+The PiCar-X is a toy car, see \autoref{fig:picarx-pt}, with all sensors and actuators connected to a RaspberryPi. Two direct current motors (DC motors) are used to move the car. A servo motor at the front is used to steer the car. The steering is a typical Ackermann steering [@Veneri:2020] known from common cars. The PiCar-X also includes grey-scale sensors for line following, infrared sensors for collision avoidance, and a camera. However, in the current example only the DC motors and the servo motor for steering are included, so far. All software components are implemented using the middleware ROS and are containerized using Docker. The ARCHES Digital Twin Framework establishes the digital thread between the physical twin and the digital shadow/twin [@barbiepicarx:2024].
 
 ![The Picar-X by SunFounder.\label{fig:picarx-pt}](./img/picarx-pt.jpg){ width=70% }
 
@@ -41,14 +43,15 @@ Lacking official CAD files for the PiCar-X, we utilized a simplified CAD model o
 
 ![The CAD model used for the PiCar-X digital model in a GAZEBO simulation.\label{fig:picarx-dm}](./img/picarx-dm.jpg){ width=70% }
 
-The PiCar-X can also be used to explore the digital twin prototype approach utilized in the smart farming application SilageControl [@silagecontrol:2024].
+In this project, we provide Docker compose files that can be used to run the software with either the digital model, digital shadow, or digital twin. Our approach to implementing a digital twin differs from others by reusing as many software components from the physical twin as possible for the digital twin [@barbiepicarx:2024].
 
 # The Digital Twin Prototype
+Developing a physical twin typically requires connecting the hardware to a development environment. However, in this setup, only one person can use the hardware at a time. For a team of engineers, this means either everyone needs their own PiCar-X or they must take turns, which can become costly, especially in real-world applications like full-scale vehicles. This is why we developed the digital twin prototype.
[…]
