### 2024-09-28 cau-se/ARCHES-PiCar-X#9 @AlexanderBarbie
enhance the paper
Based on the comments of @AlexanderFabisch I already made modification to the paper but opened this issue to track the commits better.

>Software paper
Summary:
The summary could be improved: referring to

>>This paper presents a digital twin prototype of...

>What is a digital twin and why is it useful?

>>Our goal is to provide researchers and practitioners with an affordable and straightforward example to explore various concepts of digital twins.

>What are these concepts of digital twins? In which way can they be explored?

>>Moreover, this example can be used to explore all the concepts implemented in a previous project named ARCHES (Autonomous Robotic Networks to help Human Societies), which focused on monitoring and operating an underwater network of ocean observation systems

>Again, what are these concepts and how are we able to explore these with the picar?

>>One of the outcomes of this project was the ARCHES Digital Twin Framework

>Can you briefly describe what the ARCHES Digital Twin Framework does and why it is needed?

>A statement of need:
>>Digital twins are becoming increasingly relevant...

>Why are they becoming more relevant? Certainly they are useful in a way, but it is not clear to me.

>>Industrial Internet of Things and Industry 4.0

>I think these terms need references.

>Quality of writing:
>I think the structure could be improved. Here are two suggestions.
As I pointed out before it would be useful to point out which are the digital twin concepts from the ARCHES framework that should be demonstrated by this software and then point out how this is achieved in the following sections.
>The description of the Picar, CAD model, and Digital Twin could be one section. If the automated testing pipeline is one of the concepts to be demonstrated in the paper, then I would put it in another section with all the other concepts that the software is supposed to demonstrate.

https://github.com/openjournals/joss-reviews/issues/7179
