### 2024-09-28 37c8590
merged digital twin prototype and cicd sections #9
