### 2024-09-29 b8f8892
enabled changing the i2c register address via variables, removed license text from files, adjusted compose files
