+++ PiCar-X/ros/emulators/clutchgear/nodes/clutchgear_emulator.py
@@ -1,17 +1,4 @@
 #!/usr/bin/env python3
-# Copyright 2022 - 2024 Alexander Barbie
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 
 from typing import Union
 import rospy
@@ -37,6 +24,8 @@ class Options(object):
             "pwm_pin", type=str, help="The interval in which a line in the file should be read.")
         parser.add_argument(
             "i2c_port", type=str, help="The interval in which a line in the file should be read.")
+        parser.add_argument(
+            "i2c_address", type=int, default=20,  help="The register address on the I2C.")
 
         self.args = parser.parse_args(argv)
 
@@ -64,8 +53,9 @@ class ClutchGearEmulator(AbstractClutchGearEmulator):
     :type frequency: int, optional
     """
 
-    def __init__(self, name: str, pwm_pin: str, i2c_port: str, frequency: int = 50):
-        super(ClutchGearEmulator, self).__init__(pwm_pin, i2c_port, frequency)
+    def __init__(self, name: str, pwm_pin: str, i2c_port: str, frequency: int = 50, address: int = 20):
+        super(ClutchGearEmulator, self).__init__(
+            pwm_pin, i2c_port, frequency, address)
         self.name = name
         self.frequency = frequency
         self.controller_publisher = None
@@ -106,7 +96,7 @@ if __name__ == '__main__':
     try:
         options = Options(rospy.myargv()[1:])
         clutchgear_emulator = ClutchGearEmulator(
-            options.args.name, options.args.pwm_pin, options.args.i2c_port)
+            options.args.name, options.args.pwm_pin, options.args.i2c_port, options.args.i2c_address)
         clutchgear_emulator.start()
     except rospy.ROSInterruptException:
         pass
