### 2024-09-29 f56dd61
reverted mqtt config changes and added env variable instead of hardcoded hostname/ip #8
