+++ PiCar-X/docker-compose-dtp.yml
@@ -39,7 +39,7 @@ services:
     env_file:
       - ./env/picarx.env
     tty: true
-    command: /bin/bash -c "roslaunch arches_mqtt_bridge pt_client.launch host:=mqtt-server"
+    command: /bin/bash -c "roslaunch arches_mqtt_bridge pt_client.launch host:=$${MQTT_HOST:-192.168.1.2}"
     restart: on-failure
     depends_on:
       picarx:
