+++ PiCar-X/docker-compose-dtsim.yml
@@ -48,7 +48,7 @@ services:
     env_file:
       - ./env/picarx-dt.env
     tty: true
-    command: /bin/bash -c "roslaunch arches_mqtt_bridge dt_client.launch host:=mqtt-server"
+    command: /bin/bash -c "roslaunch arches_mqtt_bridge dt_client.launch host:=$${MQTT_HOST:-192.168.1.2}"
     depends_on:
       picarx-dt:
         condition: service_healthy
