### 2024-09-29 02f41f0
added compose file to enable the start of dtp and dt on same machine #8
