+++ PiCar-X/docker-compose-dtsim-same-system.yml
@@ -1,0 +1,246 @@
+services:
+
+  mosquitto:
+    container_name: mqtt-server
+    image: eclipse-mosquitto
+    ports:
+      - 1883:1883
+    volumes:
+      - ./config/mqtt:/root/catkin_ws/mosquitto
+    command: mosquitto -c /root/catkin_ws/mosquitto/mosquitto.conf -p 1883
+
+  picarx-dt:
+    build: .
+    image: ghcr.io/cau-se/arches-picar-x/picarx:${TAG:-latest}
+    environment:
+      - ROS_HOSTNAME=picarx-dt
+      - ROS_MASTER_URI=http://picarx-dt:11313
+      - ROS_HOST_PORT=11313
+    tty: true
+    ports:
+      - 11313:11313
+    command: roscore -p 11313
+    healthcheck:
+      test: [ "CMD-SHELL", "curl $${ROS_MASTER_URI} || exit 1" ]
+      interval: 5s
+      timeout: 1s
+      retries: 5
+      start_period: 1s
+
+  shadow-dt:
+    image: ghcr.io/cau-se/arches-picar-x/picarx:${TAG:-latest}
+    env_file:
+      - ./env/picarx-dt.env
+    tty: true
+    command: /bin/bash -c "roslaunch arches_core digitalshadow_dt.launch"
+    depends_on:
+      picarx-dt:
+        condition: service_healthy
+
+  mqtt-dt:
+    image: ghcr.io/cau-se/arches-picar-x/arches-mqtt-bridge:${TAG:-latest}
+    build:
+      dockerfile: Dockerfile-arches-mqtt-bridge
+      context: ./ros
+      args:
+        - TAG=${TAG:-latest}
+        - ARCH=${ARCH:-}
+    env_file:
+      - ./env/picarx-dt.env
+    tty: true
+    command: /bin/bash -c "roslaunch arches_mqtt_bridge dt_client.launch host:=$${MQTT_HOST:-192.168.1.2}"
+    depends_on:
+      picarx-dt:
+        condition: service_healthy
+
+  motor_left-dt:
+    build:
+      dockerfile: Dockerfile
+      context: ./ros/drivers/dcmotor/
+      args:
+        - TAG=${TAG:-latest}
+        - ARCH=${ARCH:-}
+    image: ghcr.io/cau-se/arches-picar-x/drivers/dcmotor:${TAG:-latest}
+    env_file:
+      - ./env/picarx-dt.env
+    tty: true
+    privileged: true
+    volumes:
+      - /sys/class/gpio:/sys/class/gpio
+      - ${I2C:-/dev/i2c-0}:${I2C:-/dev/i2c-0}
+      - ./core:/root/catkin_ws/src/core
+      - ./ros/drivers/dcmotor:/root/catkin_ws/src/drivers/dcmotor
+    command: /bin/bash -c "roslaunch picarx_dcmotor_driver dcmotor_left.launch motor1_i2c_port:=$${I2C:-/dev/i2c-0} i2c_address:=21 motor1_direction_pin:=13"
+    depends_on:
+      picarx-dt:
+        condition: service_healthy
+
+  motor_right-dt:
+    image: ghcr.io/cau-se/arches-picar-x/drivers/dcmotor:${TAG:-latest}
+    tty: true
+    env_file:
+      - ./env/picarx-dt.env
+    privileged: true
+    volumes:
+      - /sys/class/gpio:/sys/class/gpio
+      - ${I2C:-/dev/i2c-0}:${I2C:-/dev/i2c-0}
+      - ./core:/root/catkin_ws/src/core
+      - ./ros/drivers/dcmotor:/root/catkin_ws/src/drivers/dcmotor
[…]
