### 2024-09-29 cau-se/ARCHES-PiCar-X#8 @AlexanderBarbie
sooo... I played around a bit more and added an additional docker compose file to enable the execution of the digital twin prototype and a digital twin as counterpart with running gazebo for both. Furthermore, I reverted changes on the mosquitto config, sometimes the mqtt server did not start properly on my windows system. Moreover, I added an environment variable to set the mqtt server ip, this should be easier than setting it in the compose file. I will adjust the documentation today when I am back from work.
