### 2024-09-30 cau-se/ARCHES-PiCar-X#10 @AlexanderBarbie
more documentation
Based on the comments of @AlexanderFabisch I already made modification to the documentation but opened this issue to track the commits better.

> Documentation
A statement of need: There is no statement of need in the documentation. Neither in the readme nor in the user guide.
Example usage: I am not entirely sure here. There are some examples of how to use it with a physical picar, but I cannot test it. There are also examples of how to test the digital twin in an automated test pipeline, which I could test. As I understand that is one of the main applications of the digital twin. Is this correct?
Functionality documentation: There is no API documentation. I am not sure if an API documentation is needed, but there should certainly be some additional documentation on how to work with the software apart from "run these docker commands". It could just be a link to a script with some documentation in the comments or some user guide.
Community guidelines: There is only an empty section "Contributing" in the readme.

https://github.com/openjournals/joss-reviews/issues/7179
