### 2024-09-30 19e5ce6
refactored explore and digital twin pages, more content to do #8 # 10
