+++ docs/digitaltwins/dt.md
@@ -1,0 +1,31 @@
+---
+title: The Digital Twin
+has_children: false
+nav_order: 6
+parent: Digital Twin Concept
+---
+
+**This section is based on the publications:**
+>[1] Barbie, A., & Hasselbring, W. (2024). From Digital Twins to Digital Twin Prototypes: Concepts, Formalization, and Applications. IEEE Access. [https://doi.org/10.1109/access.2024.3406510](https://doi.org/10.1109/access.2024.3406510)
+
+>[2] Barbie, A., & Hasselbring, W. (2024). Toward Reproducibility of Digital Twin Research: Exemplified with the PiCar-X. arXiv preprint arXiv:2408.13866. [https://doi.org/10.48550/ARXIV.2408.13866](https://doi.org/10.48550/ARXIV.2408.13866)
+
+# Context
+
+
+
+
+# Definition
+
+A **digital shadow** is the sum of all the data that are gathered by an embedded system from sensing, processing, or actuating. The connection from a physical twin to its digital shadow is automated. Changes on the physical twin are reflected to the digital shadow automatically. Vice versa, the digital shadow does not change the state of the physical twin.
+
+## Key Points:
+- Automatically updates based on the physical twin’s data.
+- One-way communication: physical twin data is reflected in the digital shadow.
+- In the PiCar-X setup, the **digital shadow** monitors the car's data, such as motor speed and steering angles, but does not alter its behavior.
+
+---
+
+# Digital Shadow of the PiCar-X
+
+
