+++ docs/explore/overview.md
@@ -1,0 +1,13 @@
+---
+title: Explore
+has_children: true
+nav_order: 5
+---
+
+# Explore Digital Twin Concepts
+
+
+# Explore
+* [Example for connecting physical twin and digital twin]({{ site.baseurl }}{% link digitaltwins/ptdt.md %})
+* [Example for connceting physical twin and digital shadow]({{ site.baseurl }}{% link digitaltwins/ptds.md %})
+* [Example for replacing the physical twin with a digital twin prototype]({{ site.baseurl }}{% link digitaltwins/dtp.md %})
