+++ docs/explore/adtf.md
@@ -1,0 +1,6 @@
+---
+title: The ARCHES Digital Twin Framework
+has_children: false
+nav_order: 4
+parent: Explore
+---
