+++ docs/explore/overview.md
@@ -8,6 +8,6 @@ nav_order: 5
 
 
 # Explore
-* [Example for connecting physical twin and digital twin]({{ site.baseurl }}{% link digitaltwins/ptdt.md %})
-* [Example for connceting physical twin and digital shadow]({{ site.baseurl }}{% link digitaltwins/ptds.md %})
-* [Example for replacing the physical twin with a digital twin prototype]({{ site.baseurl }}{% link digitaltwins/dtp.md %})
+* [Example for connecting physical twin and digital twin]({{ site.baseurl }}{% link explore/ptdt.md %})
+* [Example for connceting physical twin and digital shadow]({{ site.baseurl }}{% link explore/ptds.md %})
+* [Example for replacing the physical twin with a digital twin prototype]({{ site.baseurl }}{% link explore/dtps/dtp.md %})
