### 2024-09-30 829954b
fixed i2cdetect format on troubleshooting page #10
