### 2024-10-04 9a0431e
updated the digital twin part #10
