+++ docs/explore/overview.md
@@ -5,9 +5,3 @@ nav_order: 5
 ---
 
 # Explore Digital Twin Concepts
-
-
-# Explore
-* [Example for connecting physical twin and digital twin]({{ site.baseurl }}{% link explore/ptdt.md %})
-* [Example for connceting physical twin and digital shadow]({{ site.baseurl }}{% link explore/ptds.md %})
-* [Example for replacing the physical twin with a digital twin prototype]({{ site.baseurl }}{% link explore/dtps/dtp.md %})
