+++ docs/explore/adtf.md
@@ -6,21 +6,52 @@ parent: Explore
 ---
 
 # The ARCHES Digital Twin Framework
+This page gives a short overview of the ARCHES Digital Twin Framework [1]. This Framework was developed during the project ARCHES (Autonomous Robotic Networks to Help Modern Societies), where a network of underwater ocean observation systems was developed and successfully tested.
 
-# ARCHESPublisher and ARCHESSubscriber extend ROS Publishers and Subscribers
+Consider a simple stationary observation system equipped with several sensors that store data measurement probes onboard, such as hourly readings. This system is designed to allow easy addition, removal, or replacement of sensors without affecting other parts of the system. However, limitations arise when a researcher or other entity needs to access or process the sensor data in real-time. Accessing and retrieving sensor data requires an additional system with access to the sensors, which must be adapted whenever sensors are added or removed. Consequently, with each new requirement, the complexity of the observation system increases [2].
 
-![ARCHES Publisher and Subscriber](../assets/images/adtf-pubsubs.png "UML ARCHES Digital Twin Framework publisher and subscriber") 
+Observation systems, cyber-physical systems, or embedded systems in general can be viewed as distributed systems. Instead of business workflows spanning different data centers, these systems involve sensors and actuators as distinct processes that need integration into a modular core system. The objective is to develop an embedded software system where changing sensors or actuators does not impact other software components.
 
-# The Role of Skills in the ADTF
+The longstanding practice of developing embedded software using monolithic architectures has significantly influenced existing middleware and robotic systems. In these architectures, device drivers are often tightly coupled with the middleware used during the development of the embedded system~. This tight coupling creates challenges when attempting to generalize or adapt the software to other systems. For instance, when switching to a different middleware, the device driver must be redeveloped, even if the hardware logic remains unchanged. The same issue arises with algorithms that perform general tasks independent of the specific functions of the embedded system, such as pathfinding or image recognition.
+
+We decided to use the middleware [Robot Operating System (ROS)](https://ros.org), which boasts a large and active community, a clear roadmap for long-term support (LTS) releases, bindings for all popular programming languages, and a straightforward publish-subscribe mechanism. It is important to note that when referring to ROS in this context, we mean ROS1. A successor, ROS2, is being developed concurrently and is set to replace ROS1.
+
+ROS embodies five core concepts:
+
+1. **Free and Open-Source:** ROS is released under the BSD license, making it freely available and open-source.
+
+2. **Peer-to-Peer:** Instead of relying on a central server, ROS-based embedded software systems consist of multiple processes connected at runtime in a peer-to-peer topology. The message broker facilitating this communication is called the ROS Master.
+
+3. **Tool-Based:** ROS features a microkernel design. Rather than using a monolithic structure with a master module providing all functions (such as a build system or global clock), ROS employs various tools to build and run different components. This approach reduces the complexity of the core library and provides developers with significant freedom in their choice of tools and libraries.
+
+4. **Multi-Lingual:** ROS is designed to be language-neutral, with its specification focused on the messaging layer. Connection negotiation and configuration between processes occur via XML-RPC. A language-neutral interface definition language describes the messages sent between modules, enabling cross-language development.
+
[…]
