### 2024-10-06 4e9dcdf
refactored for better digital shadow example and fixed minor errors
