+++ PiCar-X/ros/skills/ackermann_drive/src/picarx_ackermann_drive/monitor.py
@@ -1,0 +1,160 @@
+#!/usr/bin/env python3
+
+from typing import Union
+import rospy
+from arches_core.skills import Skill
+from picarx_msgs.msg import DriveStatus
+import argparse
+
+
+class AckermannStartOptions(object):
+    """Start argument parser to the digital shadow skill.
+        You have to set the name and a unique id. If you leave out a unique id,
+        an autoincrement value will be used as id.
+
+        Note:
+            If you use Docker, we recommand to set the ids.
+    """
+
+    def __init__(self, argv):
+        """
+        Initialize the argument parser with command line arguments.
+
+        Args:
+            argv (list): List of command line arguments.
+        """
+        parser = argparse.ArgumentParser()
+        parser.add_argument(
+            "name", type=str, help="The name of the digital shadow skill")
+        parser.add_argument(
+            "uid", type=str, help="The unique identifier of the digital shadow skill. Leave blank if you want and automated generation.")
+
+        self.args = parser.parse_args(argv)
+
+
+class AckermannMonitorSkill(Skill):
+    """
+    AckermannDriveSkill class for controlling the Ackermann drive.
+
+    Attributes:
+        motor_left_publisher (rospy.Publisher): Publisher for the left motor.
+        motor_right_publisher (rospy.Publisher): Publisher for the right motor.
+        clutchgeer_publisher (rospy.Publisher): Publisher for the clutch gear.
+    """
+
+    def __init__(self, name: str, uid: str = None) -> None:
+        """
+        Initialize the AckermannDriveSkill.
+
+        Args:
+            name (str): Name of the skill.
+            uid (str, optional): Unique identifier of the skill. Defaults to None.
+        """
+        super(AckermannMonitorSkill, self).__init__(name, 'Monitoring', uid)
+        self.motor_left_publisher = None
+        self.motor_right_publisher = None
+        self.clutchgeer_publisher = None
+
+    @property
+    def motor_left_publisher(self) -> Union[None, rospy.Publisher]:
+        """
+        Get the left motor publisher.
+
+        Returns:
+            Union[None, rospy.Publisher]: The left motor publisher.
+        """
+        return self.__motor_left_publisher
+
+    @motor_left_publisher.setter
+    def motor_left_publisher(self, publisher: Union[None, rospy.Publisher]) -> Union[None, rospy.Publisher]:
+        """
+        Set the left motor publisher.
+
+        Args:
+            publisher (Union[None, rospy.Publisher]): The left motor publisher.
+
+        Raises:
+            ValueError: If the publisher is not of type rospy.Publisher or None.
+        """
+        if publisher is None:
+            self.__motor_left_publisher = None
+        elif isinstance(publisher, rospy.Publisher):
+            self.__motor_left_publisher = publisher
+        else:
+            raise ValueError(
+                "Publisher has to be of type rospy.Publisher or None, but {} was given.".format(type(publisher)))
+
+    @property
+    def motor_right_publisher(self) -> Union[None, rospy.Publisher]:
+        """
+        Get the right motor publisher.
+
+        Returns:
+            Union[None, rospy.Publisher]: The right motor publisher.
+        """
+        return self.__motor_right_publisher
+
+    @motor_right_publisher.setter
+    def motor_right_publisher(self, publisher: Union[None, rospy.Publisher]) -> Union[None, rospy.Publisher]:
+        """
+        Set the right motor publisher.
+
+        Args:
+            publisher (Union[None, rospy.Publisher]): The right motor publisher.
+
+        Raises:
+            ValueError: If the publisher is not of type rospy.Publisher or None.
+        """
+        if publisher is None:
+            self.__motor_right_publisher = None
[…]
