+++ docs/explore/dtps/dtnosim.md
@@ -18,17 +18,17 @@ There are three environment files with relevant variables:
 
 ## Start the Digital Twin without GAZEBO
 ```console
-# Build and execute the Docker Containers
+# Build Docker Containers if not previously done
 docker compose -f docker-compose-core.yml build 
 docker compose -f docker-compose-dt.yml build 
+
+# Start the DT
 docker compose -f docker-compose-dt.yml up
 ```
 
 ## Start the Digital Twin Prototype with GAZEBO
 ```console
-# Build and execute the Docker Containers
-docker compose -f docker-compose-core.yml build 
-docker compose -f docker-compose-dtp.yml build 
+# Start the DTP
 docker compose -f docker-compose-dtp.yml up
 ```
 
