### 2024-10-06 03d1711
added some links to used technologies etc. on contribution page #10
