+++ docs/installation/installation_macos.md
@@ -7,6 +7,3 @@ nav_order: 3
 
 # Installation on MacOs
 This example will not work on MacOs, since relevant kernel modules are missing (i2c-stub and gpio-mockup).
-
-
-- The PiCar-X folder contains a Digital Twin Prototype, developed using the ARCHES Digital Twin Framework. This prototype is accompanied by various Docker compose files, each illustrating different aspects of the digital twin concept. **<span style="color:red">This project moved to a [separate repository](https://github.com/cau-se/ARCHES-PiCar-X).</span>**
