### 2024-10-07 9120e03
trying to answer the how to explore comment #9
