+++ paper/paper.md
@@ -29,7 +29,7 @@ This paper presents a digital twin prototype of the [PiCar-X by Sunfounder](http
 Digital twins are becoming increasingly relevant in the the Industrial Internet of Things and Industry 4.0 [@kritzinger:2018], as they enhance existing capabilities for monitoring and controlling cyber-physical systems. This is achieved by integrating a digital representation of the real system in the form of a digital model. However, the concept of digital twins lacks a consensual definition and faces validation challenges, partly due to the scarcity of reproducible modules or source codes in existing studies. While many applications are described in case studies, they often lack detailed, re-usable specifications for researchers and engineers. This can lead to confusions, since modern simulations or enhanced climate models are also often praised as digital twins [@barbie:2024].
 
 In [@barbie:2024], we formally specified a digital twin concept including its sub-concepts physical twin, digital model, digital template, digital thread, digital shadow, digital twin, and digital twin prototype using the Object-Z notation and they are the basis for this paper and the PiCar-X example. These concepts were developed for a network of ocean observation systems and the results were evaluated in a real-world mission in the Baltic sea in Oktober 2020 [@demomission:2020]. One of the results of the successful proof-of-concept was the ARCHES Digital Twin Framework [@ADTF:2022], a software package providing the functionality to implement the digital thread between physical twins and digital twins for data and command exchange.
-Ocean observation systems use quite specific and expensive hardware, hence, we see the need of a cheap lab experiemnt to enable independent evaluation and exploration of the different concepts. The PiCar-X example demonstrates all the concepts from the ocean observation system and includes also a full integration test pipeline, see our GitHub Repository [@archespicarx:2024]. In [@barbiepicarx:2024], we elaborate in more detail how the PiCar-X can be used to evaluate all these concepts and how they differ from each other in possible code implementations.
+Ocean observation systems use quite specific and expensive hardware, hence, we see the need of a cheap lab experiemnt to enable independent evaluation and exploration of the different concepts. The PiCar-X example [@archespicarx:2024] is implemented based on ROS and the ARCHES Digital Twin Framework and  demonstrates how the framework was used for ocean observation system. Docker Compose files facilitate the launching of the physical twin, digital model, digital shadow, digital twin, and digital twin prototype, which not only differ in the complexity of the compose files but also source code. The digital twin prototype can also be used to connect to the digital shadow or digital twin without the physical PiCar-X, enabling exploration of the communication between them and how the binary messages exchange works. Moreover, our example includes an integration test pipeline that is similar to the appraoch in ARCHES. In [@barbiepicarx:2024], we elaborate in more detail how the PiCar-X can be used to evaluate all these concepts and how they differ from each other in possible code implementations.
 
 
 # The PiCar-X and its Digital Twin
@@ -43,7 +43,7 @@ Lacking official CAD files for the PiCar-X, we utilized a simplified CAD model o
 
 ![The CAD model used for the PiCar-X digital model in a GAZEBO simulation.\label{fig:picarx-dm}](./img/picarx-dm.jpg){ width=70% }
 
-We provide Docker compose files that can be used to run the software with either the digital shadow, digital twin, or digital twin prototype.
+We provide Docker compose files that can be used to run the software with either the digital model, digital shadow, digital twin, or digital twin prototype.
 
 # The Digital Twin Prototype for Development and Automated Integration Testing in CI/CD Pipelines
 Developing a physical twin typically requires connecting the hardware to a development environment. However, in such a setup, only one person can use the hardware at a time. For a team of engineers, this means either everyone needs their own PiCar-X or they must take turns, which can become costly, especially in real-world applications like full-scale vehicles.
