### 2024-10-07 cau-se/ARCHES-PiCar-X#9 @AlexanderBarbie
@AlexanderFabisch sorry, took me a while because I have a lot of stuff to do.. but I finally revised the paper and tried to make it more concrete in the areas you highlighted. However, I'm already over the 1,000-word limit and would prefer not to write a third paper with the same definitions as in the two referenced ones. Instead, I've added more about the digital twin concept to the documentation.
