### 2024-10-07 #7179 @AlexanderBarbie
> @AlexanderBarbie - Can you update us on where things are? (I know there are lots of issues, but perhaps you can summarize this here?)

I'm sorry, I did not see you last comment.

I opened two issues to work on the comments of @AlexanderFabisch.

The first one to [revise the paper](https://github.com/cau-se/ARCHES-PiCar-X/issues/9), which I did today. In short: I tried to be more specific and what the PiCar-X can be used for. Since there is already a publication with regards to the formalization of the used digital twin concept and also a preprint which describes in more detail how the different concepts (digital mode, shadow, twin, etc.) could be implemented with the PiCar-X, I opened as second Issue.

The second issue was to [enhance the documentation](https://github.com/cau-se/ARCHES-PiCar-X/issues/10). Here I briefly introduced the different concept with small examples and added a view figures. Furthermore I added a description of the key features of the ARCHES Digital Twin Framework, which is used by the PiCar-X to exchange data/commands between physical and digital twin. This issue also affects the issue opened by @mrsonandrade. [He found a bug](https://github.com/cau-se/ARCHES-PiCar-X/issues/8) which I (hopefully) solved, but I am waiting whether it is working for him now. I will make a few small changes to the documentation over the weekend, but overall, I think it's pretty much finished, if there are no new comments/suggestions.
