### 2024-10-08 #7179 @AlexanderFabisch
I checked all the points in my list. Everything looks really good from my perspective. The documentation is much more detailed and useful now. The paper is easier to read.
