### 2024-10-08 cau-se/ARCHES-PiCar-X#9 @AlexanderFabisch
Looks very good!

Just one more comment:  There is a typo in "possible sources of erros" -> errors

(I'd suggest to use a spell checker like aspell on the file to remove all typos that I missed.)
