### 2024-10-08 2eccadb
fixed typos #9 #10
