+++ docs/digitaltwins/dt.md
@@ -40,5 +40,5 @@ This example has no separate docker compose file to be executed, as we enhance t
 
 
 
-# Futher References
+# Further References
 >[3] Kritzinger, W., Karner, M., Traar, G., Henjes, J., & Sihn, W. (2018). Digital Twin in manufacturing: A categorical literature review and classification. Ifac-PapersOnline, 51(11), 1016-1022. [https://doi.org/10.1016/j.ifacol.2018.08.474](https://doi.org/10.1016/j.ifacol.2018.08.474)
