### 2024-10-08 cau-se/ARCHES-PiCar-X#9 @AlexanderBarbie
great suggestion, don't know why I haven't thought about a spell checker in vscode before..  found quite a lot!
