### 2024-10-08 #7179 @AlexanderBarbie
@danielskatz I fixed some typos in the paper
