### 2024-10-09 cau-se/ARCHES-PiCar-X#8 @mrsonandrade
@danielskatz sorry for the late reply.

@AlexanderBarbie thank you for the effort! I have two points:
- When I click on “Explore” at the end of the “Getting Started” page, it goes to the “Digital Twin Concept” page (not to the Explore one, as expected).
- By following the new steps [here](https://cau-se.github.io/ARCHES-PiCar-X/explore/dtps/dtpdt.html#start-digital-twin-prototype-and-digital-twin-with-separate-gazebo-simulation), when I try to register the additional address space (0x15) by using `sudo modprobe i2c-stub chip_addr=[0x14, 0x15]`, I get the following: `modprobe: ERROR: could not insert ‘i2c_stub’: Invalid argument`. The command works fine if I have only one address, e.g. `sudo modprobe i2c-stub chip_addr=0x14`. What should I do?
