### 2024-10-09 5975a4a
fixed wrong link on Getting Started page and fixed wrong i2c-stub command to create two i2c registers #8
