### 2024-10-09 cau-se/ARCHES-PiCar-X#8 @AlexanderBarbie
please try `sudo modprobe i2c-stub chip_addr=0x14,0x15`
I would add a list of values in general with [], but looks like i2c-stub wants a list of values comma separated.
