### 2024-10-11 cau-se/ARCHES-PiCar-X#8 @mrsonandrade
Great! Everything is working now!
