### 2024-10-11 #7179 @mrsonandrade
I’m almost done with all the points on my list. My last point concerns the **State of the field**, which should be in the paper. Thus, are there other related packages? If yes, please, compare them. If not, I think it is valid to clarify this in the paper.
