+++ paper/paper.md
@@ -31,6 +31,8 @@ Digital twins are becoming increasingly relevant in the Industrial Internet of T
 In [@barbie:2024], we formally specified a digital twin concept including its sub-concepts physical twin, digital model, digital template, digital thread, digital shadow, digital twin, and digital twin prototype using the Object-Z notation and they are the basis for this paper and the PiCar-X example. These concepts were developed for a network of ocean observation systems and the results were evaluated in a real-world mission in the Baltic sea in October 2020 [@demomission:2020]. One of the results of the successful proof-of-concept was the ARCHES Digital Twin Framework [@ADTF:2022], a software package providing the functionality to implement the digital thread between physical twins and digital twins for data and command exchange.
 Ocean observation systems use quite specific and expensive hardware, hence, we see the need of a cheap lab experiment to enable independent evaluation and exploration of the different concepts. The PiCar-X example [@archespicarx:2024] is implemented based on ROS and the ARCHES Digital Twin Framework and  demonstrates how the framework was used for ocean observation system. Docker Compose files facilitate the launching of the physical twin, digital model, digital shadow, digital twin, and digital twin prototype, which not only differ in the complexity of the compose files but also source code. The digital twin prototype can also be used to connect to the digital shadow or digital twin without the physical PiCar-X, enabling exploration of the communication between them and how the binary messages exchange works. Moreover, our example includes an integration test pipeline that is similar to the approach in ARCHES. In [@barbiepicarx:2024], we elaborate in more detail how the PiCar-X can be used to evaluate all these concepts and how they differ from each other in possible code implementations.
 
+# Related Work
+Examples of digital twin implementations based on sensor recordings or simulation data already exist, with some even providing a Docker setup [@Eckhart:2018; @Russo:2023]. @Fogli:2023 offer an abstract example involving chaos testing, which includes integration tests. However, none of these publications present an implementation of a digital twin that could be easily connected with the proposed physical twin. At the time the ARCHES PiCar-X was published, and to the best of our knowledge, no similar software packages were available that demonstrated the fundamental implementation of the various concepts from physical twin to digital twin (prototype), including automated CI/CD pipelines with actual integration tests.
 
 # The PiCar-X and its Digital Twin
 Lacking a consensual definition of a digital twin, the range of interpretations spans from a complex simulation to a completely mirrored status of the physical device in real-time. In our definition [@barbie:2024], the digital twin is connected to the physical twin over the entire life cycle for automated bidirectional data exchange, i.e. changes made to the digital twin lead to adapted behavior of the physical twin and vice-versa. The implementation of this capabilties can vary, with model driven approaches are being very popular [@barbie:2024]. To reduce implementation complexity and possible sources of errors, our approach for the implementation of digital twins differs from others by reusing as many software components from the physical twin as possible [@barbiepicarx:2024].
