+++ paper/paper.bib
@@ -100,3 +100,38 @@ url = {https://github.com/acmsigsoft/EmpiricalStandards}
   title = {{Digital Twin in manufacturing: A categorical literature review and classification}},
   journal = {{IFAC}-{PapersOnLine}}
 }
+
+@inproceedings{Eckhart:2018,
+  title = {{A Specification-based State Replication Approach for Digital Twins}},
+  DOI = {10.1145/3264888.3264892},
+  booktitle = {Proceedings of the 2018 Workshop on Cyber-Physical Systems Security and PrivaCy},
+  publisher = {ACM},
+  author = {Eckhart, Matthias and Ekelhart, Andreas},
+  year = {2018},
+  month = jan,
+}
+
+@article{Russo:2023,
+  title = {{LiDiTE: A Full-Fledged and Featherweight Digital Twin Framework}},
+  volume = {20},
+  ISSN = {2160-9209},
+  DOI = {10.1109/tdsc.2023.3236798},
+  number = {6},
+  journal = {IEEE Transactions on Dependable and Secure Computing},
+  publisher = {IEEE},
+  author = {Russo,  Enrico and Costa,  Gabriele and Longo,  Giacomo and Armando,  Alessandro and Merlo,  Alessio},
+  year = {2023},
+  month = nov,
+  pages = {4899–4912}
+}
+
+@article{Fogli:2023,
+  title = {{Chaos Engineering for Resilience Assessment of Digital Twins}},
+  ISSN = {1941-0050},
+  DOI = {10.1109/tii.2023.3264101},
+  journal = {IEEE Transactions on Industrial Informatics},
+  publisher = {IEEE},
+  author = {Fogli, Mattia and Giannelli, Carlo and Poltronieri, Filippo and Stefanelli, Cesare and Tortonesi,  Mauro},
+  year = {2023},
+  pages = {1–9}
+}
