### 2024-10-12 cau-se/ARCHES-PiCar-X#9 @AlexanderBarbie
@mrsonandrade I reopened this issue for discussion.

>I’m almost done with all the points on my list. My last point concerns the State of the field, which should be in the paper. Thus, are there other related packages? If yes, please, compare them. If not, I think it is valid to clarify this in the paper.

I added a related work section, but since I already exceeded the 1000 words, I only briefly discuss the differences. Is this okay for you?
