### 2024-10-13 cau-se/ARCHES-PiCar-X#9 @mrsonandrade
@AlexanderBarbie - Excellent!
