### 2024-10-14 cae34f0
minor changes in paper.md
