### 2024-10-14 #7179 @danielskatz
@AlexanderBarbie - I'm suggesting a bunch of small changes in https://github.com/cau-se/ARCHES-PiCar-X/pull/12 - please merge this, or let me know what you disagree with, then we can continue to acceptance and publication
