### 2024-10-14 05bd450
minor changes for JOSS publication (#12)
