+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{barbie:2024,
-  title = {{From Digital Twins to Digital Twin Prototypes: Concepts, Formalization, and Applications}},
+  title = {From Digital Twins to Digital Twin Prototypes: Concepts, Formalization, and Applications},
   volume = {12},
   ISSN = {2169-3536},
   DOI = {10.1109/access.2024.3406510},
@@ -11,7 +11,7 @@
 }
 
 @article{silagecontrol:2024,
-  title = {{Digital Twin Prototypes for Supporting Automated Integration Testing of Smart Farming Applications}},
+  title = {Digital Twin Prototypes for Supporting Automated Integration Testing of Smart Farming Applications},
   volume = {16},
   ISSN = {2073-8994},
   DOI = {10.3390/sym16020221},
@@ -29,14 +29,14 @@
   year = {2021},
   publisher = {Institute of Electrical and Electronics Engineers ({IEEE})},
   author = {Alexander Barbie and Niklas Pech and Wilhelm Hasselbring and Sascha Flogel and Frank Wenzhofer and Michael Walter and Elena Shchekinova and Marc Busse and Matthias Turk and Michael Hofbauer and Stefan Sommer},
-  title = {{Developing an Underwater Network of Ocean Observation Systems with Digital Twin Prototypes - A Field Report from the Baltic Sea}},
+  title = {Developing an Underwater Network of Ocean Observation Systems with Digital Twin Prototypes - A Field Report from the {B}altic {S}ea},
   journal = {{IEEE} Internet Computing}
 }
 
 @misc{ADTF:2022,
   doi = {10.3289/sw_arches_core_1.0.0},
   year = {2022},
-  publisher = {{{GEOMAR} Helmholtz Centre for Ocean Research Kiel}},
+  publisher = {{GEOMAR} {H}elmholtz {C}entre for {O}cean {R}esearch {K}iel},
   author = {Alexander Barbie and Niklas Pech},
   title = {{ARCHES Digital Twin Framework}},
   note = {[Software]}
@@ -45,7 +45,7 @@
 @ARTICLE{studyembeddedtesting:2018,  
 author={Garousi, Vahid and Felderer, Michael and Karap{\i}{\c{c}}ak, {\c{C}}a{\u{g}}r{\i} Murat and Y{\i}lmaz, U{\u{g}}ur},  
 journal={IEEE Software},   
-title={{What We Know about Testing Embedded Software}},   
+title={What We Know about Testing Embedded Software},   
 year={2018},  
 volume={35},  
 number={4},  
@@ -54,7 +54,7 @@ doi={10.1109/MS.2018.2801541}}
 
 @MISC{empiricalstandardsgithub, 
 author = {Ralph, Paul and Baltes, Sebastian and Bianculli, Domenico and Dittrich, Yvonne and Felderer, Michael and Feldt, Robert and Filieri, Antonio and Furia, Carlo and Graziotin, Daniel and He, Pinjia and Hoda, Rashina and Juristo, Natalia and Kitchenham, Barbara and Robbes, Romain and Méndez Fernández, Daniel and Molleri, Jefferson and Spinellis, Diomidis and Staron, Miroslaw and Stol, Klaas-Jan and Vegas, Sira},
-title = {{ACM SIGSOFT Empirical Standards on GitHub}},
+title = {ACM SIGSOFT Empirical Standards on {GitHub}},
 url = {https://github.com/acmsigsoft/EmpiricalStandards}
 }
 
@@ -67,7 +67,7 @@ url = {https://github.com/acmsigsoft/EmpiricalStandards}
   number = {6},
   pages = {907--927},
   author = {M. Veneri and M. Massaro},
-  title = {{The effect of Ackermann steering on the performance of race cars}},
[…]
