### 2024-10-14 #7179 @danielskatz
@AlexanderBarbie - please check this PDF to make sure it looks ok, and assuming so, please
