### 2024-10-15 71eb9bd
fixed driving backward for the dt/dtp and added bad example for digital model #11
