+++ PiCar-X/core/picarx/tests/test_pwm.py
@@ -18,8 +18,6 @@ class TestPWM(unittest.TestCase):
         self.assertEqual(self.pwm.channel, 1)
         self.assertEqual(self.pwm.debug, "critical")
         self.assertIsInstance(self.pwm.register_channel, WordRegister)
-        self.assertIsInstance(self.pwm.register_prescaler, WordRegister)
-        self.assertIsInstance(self.pwm.register_frequency, WordRegister)
 
     def test_channel_property(self):
         self.pwm.channel = "P2"
@@ -31,12 +29,6 @@ class TestPWM(unittest.TestCase):
         self.pwm.period = 1000
         self.assertEqual(self.pwm.period, 999)
 
-    def test_prescaler_property(self):
-        self.pwm.prescaler = 10
-        expect_result = to_big_endian(10)
-        self.mock_smbus_instance.write_word_data.assert_called_with(
-            SunFounderPWMValues.DEFAULT_I2C_ADRESS.value, SunFounderPWMValues.REGISTER_PRESCALER.value + 1, expect_result
-        )
 
     def test_duty_cycle_property(self):
         self.pwm.duty_cycle = 50
@@ -45,12 +37,7 @@ class TestPWM(unittest.TestCase):
             SunFounderPWMValues.REGISTER_CHANNEL.value + 1, expect_result
         )
 
-    def test_frequency_property(self):
-        self.pwm.frequency = 1000
-        expect_result = to_big_endian(1000)
-        self.mock_smbus_instance.write_word_data.assert_called_with(
-            SunFounderPWMValues.DEFAULT_I2C_ADRESS.value, SunFounderPWMValues.REGISTER_FREQUENCY.value + 1, expect_result
-        )
+
 
 if __name__ == '__main__':
     unittest.main()
