### 2024-10-15 #7179 @AlexanderBarbie
Sooo... I fixed some bugs and uploaded the released version now. I think it is under review status on zenodo but there was a DOI generated already.

version: [v.1.1.1](https://github.com/cau-se/ARCHES-PiCar-X/releases/tag/v1.1.1)
doi: 10.5281/zenodo.13937454
