### 2024-10-16 #7179 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
