### 2024-10-16 #7179 @danielskatz
Congratulations to @AlexanderBarbie (Alexander Barbie) and @whasselbring on your publication!!

And thanks to @mrsonandrade and @AlexanderFabisch for reviewing!
JOSS depends on volunteers and couldn't be successful without you
