### 2024-10-16 #7179 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
